% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/absval.cwres.vs.cov.bw.R
\name{absval.cwres.vs.cov.bw}
\alias{absval.cwres.vs.cov.bw}
\title{Absolute conditional weighted residuals vs covariates for Xpose 4}
\usage{
absval.cwres.vs.cov.bw(object, xlb = "|CWRES|", main = "Default", ...)
}
\arguments{
\item{object}{An xpose.data object.}

\item{xlb}{A string giving the label for the x-axis. \code{NULL} if none.}

\item{main}{The title of the plot.  If \code{"Default"} then a default title
is plotted. Otherwise the value should be a string like \code{"my title"} or
\code{NULL} for no plot title.  For \code{"Default"} the function
\code{\link{xpose.multiple.plot.title}} is used.}

\item{\dots}{Other arguments passed to \code{\link{xpose.plot.bw}}.}
}
\value{
Returns a stack of box-and-whisker plots of |CWRES| vs covariates.
}
\description{
This creates a stack of box and whisker plot of absolute population
conditional weighted residuals (|CWRES|) vs covariates, and is a specific
function in Xpose 4.  It is a wrapper encapsulating arguments to the
code\link{xpose.plot.bw} function. Most of the options take their default
values from xpose.data object but may be overridden by supplying them as
arguments.
}
\details{
Each of the covariates in the Xpose data object, as specified in
\code{object@Prefs@Xvardef$Covariates}, is evaluated in turn, creating a
stack of plots.

Conditional weighted residuals (CWRES) require some extra steps to
calculate. See \code{\link{compute.cwres}} for details.

A wide array of extra options controlling box-and-whisker plots are
available. See \code{\link{xpose.plot.bw}} for details.
}
\examples{

\dontrun{
## We expect to find the required NONMEM run and table files for run
## 5 in the current working directory
xpdb5 <- xpose.data(5)

## Here we load the example xpose database 
data(simpraz.xpdb)
xpdb <- simpraz.xpdb

## A vanilla plot
absval.cwres.vs.cov.bw(xpdb)


## A custom plot
absval.cwres.vs.cov.bw(xpdb, bwdotcol="white", 
  bwdotpch=15,
  bwreccol="red",
  bwrecfill="red",
  bwumbcol="red",
  bwoutpch=5,
  bwoutcol="black")
}
}
\seealso{
\code{\link{xpose.plot.bw}}, \code{\link{xpose.panel.bw}},
\code{\link{compute.cwres}}, \code{\link[lattice]{bwplot}},
\code{\link{xpose.prefs-class}}, \code{\link{xpose.data-class}}

Other specific functions: \code{\link{absval.cwres.vs.pred.by.cov}},
  \code{\link{absval.cwres.vs.pred}},
  \code{\link{absval.iwres.cwres.vs.ipred.pred}},
  \code{\link{absval.iwres.vs.cov.bw}},
  \code{\link{absval.iwres.vs.idv}},
  \code{\link{absval.iwres.wres.vs.ipred.pred}},
  \code{\link{absval_delta_vs_cov_model_comp}},
  \code{\link{autocorr.iwres}}, \code{\link{par_cov_hist}},
  \code{\link{par_cov_qq}}, \code{\link{par_cov_summary}},
  \code{\link{xpose4-package}}
}
\author{
E. Niclas Jonsson, Mats Karlsson, Andrew Hooker & Justin Wilkins
}
\keyword{methods}
