% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/absval.cwres.vs.pred.by.cov.R
\name{absval.cwres.vs.pred.by.cov}
\alias{absval.cwres.vs.pred.by.cov}
\title{Absolute value of the conditional weighted residuals vs. population
predictions, conditioned on covariates, for Xpose 4}
\usage{
absval.cwres.vs.pred.by.cov(object, ylb = "|CWRES|", type = "p",
  smooth = TRUE, idsdir = "up", main = "Default", ...)
}
\arguments{
\item{object}{An xpose.data object.}

\item{ylb}{A string giving the label for the y-axis. \code{NULL} if none.}

\item{type}{Type of plot. The default is points only ("p"), but lines ("l")
and both ("b") are also available.}

\item{smooth}{Logical value indicating whether an x-y smooth should be
superimposed. The default is TRUE.}

\item{idsdir}{Direction for displaying point labels. The default is "up",
since we are displaying absolute values.}

\item{main}{The title of the plot.  If \code{"Default"} then a default title
is plotted. Otherwise the value should be a string like \code{"my title"} or
\code{NULL} for no plot title.  For \code{"Default"} the function
\code{\link{xpose.multiple.plot.title}} is used.}

\item{\dots}{Other arguments passed to \code{link{xpose.plot.default}}.}
}
\value{
Returns a stack of xyplots of |CWRES| vs PRED, conditioned on
covariates.
}
\description{
This is a plot of absolute population conditional weighted residuals
(|CWRES|) vs population predictions (PRED) conditioned by covariates, a
specific function in Xpose 4. It is a wrapper encapsulating arguments to the
\code{xpose.plot.default} function. Most of the options take their default
values from xpose.data object but may be overridden by supplying them as
arguments.
}
\details{
Each of the covariates in the Xpose data object, as specified in
\code{object@Prefs@Xvardef$Covariates}, is evaluated in turn, creating a
stack of plots.

The \code{main} argument is not supported owing to the multiple plots
generated by the function.

Conditional weighted residuals (CWRES) require some extra steps to
calculate. See \code{\link{compute.cwres}} for details.

A wide array of extra options controlling xyplots are available. See
\code{\link{xpose.plot.default}} for details.
}
\examples{

\dontrun{
## We expect to find the required NONMEM run and table files for run
## 5 in the current working directory
xpdb <- xpose.data(5)

## Here we load the example xpose database 
data(simpraz.xpdb)
xpdb <- simpraz.xpdb

## A vanilla plot
absval.cwres.vs.pred.by.cov(xpdb)

## Custom axis labels
absval.cwres.vs.pred.by.cov(xpdb, ylb="|CWRES|", xlb="PRED")

## Custom colours and symbols, no IDs
absval.cwres.vs.pred.by.cov(xpdb, cex=0.6, pch=3, col=1, ids=FALSE)
}

}
\seealso{
\code{\link{absval.cwres.vs.pred}},
\code{\link{xpose.plot.default}}, \code{\link{xpose.panel.default}},
\code{\link[lattice]{xyplot}}, \code{\link{xpose.prefs-class}},
\code{\link{compute.cwres}}, \code{\link{xpose.data-class}}

Other specific functions: \code{\link{absval.cwres.vs.cov.bw}},
  \code{\link{absval.cwres.vs.pred}},
  \code{\link{absval.iwres.cwres.vs.ipred.pred}},
  \code{\link{absval.iwres.vs.cov.bw}},
  \code{\link{absval.iwres.vs.idv}},
  \code{\link{absval.iwres.wres.vs.ipred.pred}},
  \code{\link{absval_delta_vs_cov_model_comp}},
  \code{\link{autocorr.iwres}}, \code{\link{par_cov_hist}},
  \code{\link{par_cov_qq}}, \code{\link{par_cov_summary}},
  \code{\link{xpose4-package}}
}
\author{
E. Niclas Jonsson, Mats Karlsson, Andrew Hooker & Justin Wilkins
}
\keyword{methods}
