% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/absval.iwres.vs.cov.bw.R
\name{absval.iwres.vs.cov.bw}
\alias{absval.iwres.vs.cov.bw}
\title{box and whisker plots of the abslute value of the 
individual weighted residuals vs. covariates}
\usage{
absval.iwres.vs.cov.bw(object, xlb = "|iWRES|", main = "Default", ...)
}
\arguments{
\item{object}{An "xpose.data" object.}

\item{xlb}{A string giving the label for the x-axis. \code{NULL} if none.}

\item{main}{A string giving the plot title or \code{NULL} if none.}

\item{...}{Other arguments passed to \code{\link{xpose.panel.default}}.}
}
\value{
An xpose.multiple.plot object
}
\description{
box and whisker plots of the abslute value of the 
individual weighted residuals vs. covariates
}
\seealso{
Other specific functions: \code{\link{absval.cwres.vs.cov.bw}},
  \code{\link{absval.cwres.vs.pred.by.cov}},
  \code{\link{absval.cwres.vs.pred}},
  \code{\link{absval.iwres.cwres.vs.ipred.pred}},
  \code{\link{absval.iwres.vs.idv}},
  \code{\link{absval_delta_vs_cov_model_comp}},
  \code{\link{autocorr.iwres}}, \code{\link{cov.splom}},
  \code{\link{cwres_wres_vs_x}},
  \code{\link{ind.plots.cwres.hist}},
  \code{\link{ind.plots.cwres.qq}},
  \code{\link{par_cov_hist}}, \code{\link{par_cov_qq}},
  \code{\link{par_cov_summary}},
  \code{\link{xpose4-package}}
}
