% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/absval.iwres.vs.idv.R
\name{absval.iwres.vs.idv}
\alias{absval.iwres.vs.idv}
\title{abslute value of the 
individual weighted residuals vs. the indenpendent variable}
\usage{
absval.iwres.vs.idv(object, ylb = "|iWRES|", smooth = TRUE, idsdir = "up",
  type = "p", ...)
}
\arguments{
\item{object}{An "xpose.data" object.}

\item{ylb}{A string giving the label for the y-axis. \code{NULL} if none.}

\item{...}{Other arguments passed to \code{\link{xpose.panel.default}}.}
}
\value{
A lattice object
}
\description{
abslute value of the 
individual weighted residuals vs. the indenpendent variable
}
\seealso{
Other specific functions: \code{\link{absval.cwres.vs.cov.bw}},
  \code{\link{absval.cwres.vs.pred.by.cov}},
  \code{\link{absval.cwres.vs.pred}},
  \code{\link{absval.iwres.cwres.vs.ipred.pred}},
  \code{\link{absval.iwres.vs.cov.bw}},
  \code{\link{absval.iwres.wres.vs.ipred.pred}},
  \code{\link{absval_delta_vs_cov_model_comp}},
  \code{\link{autocorr.iwres}}, \code{\link{par_cov_hist}},
  \code{\link{par_cov_qq}}, \code{\link{par_cov_summary}},
  \code{\link{xpose4-package}}
}
