% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/absval.iwres.vs.ipred.R
\name{absval.iwres.vs.ipred}
\alias{absval.iwres.vs.ipred}
\title{Absolute individual weighted residuals vs individual predictions for Xpose 4}
\usage{
absval.iwres.vs.ipred(object, ylb = "|iWRES|", type = "p", ids = FALSE,
  idsdir = "up", smooth = TRUE, ...)
}
\arguments{
\item{object}{An xpose.data object.}

\item{ylb}{A string giving the label for the y-axis. \code{NULL} if none.}

\item{type}{Type of plot. The default is points only ("p"), but lines ("l")
and both ("b") are also available.}

\item{ids}{Should id values be displayed?}

\item{idsdir}{Direction for displaying point labels. The default is "up",
since we are displaying absolute values.}

\item{smooth}{Logical value indicating whether an x-y smooth should be
superimposed.  The default is TRUE.}

\item{\dots}{Other arguments passed to \code{link{xpose.plot.default}}.}
}
\value{
Returns an xyplot of |IWRES| vs IPRED.
}
\description{
This is a plot of absolute individual weighted residuals (|IWRES|) vs
individual predictions (IPRED), a specific function in Xpose 4. It is a
wrapper encapsulating arguments to the \code{xpose.plot.default} function.
Most of the options take their default values from xpose.data object but may
be overridden by supplying them as arguments.
}
\details{
A wide array of extra options controlling xyplots are available. See
\code{\link{xpose.plot.default}} for details.
}
\examples{

\dontrun{
## We expect to find the required NONMEM run and table files for run
## 5 in the current working directory
xpdb5 <- xpose.data(5)
}

## Here we load the example xpose database 
data(simpraz.xpdb)
xpdb <- simpraz.xpdb

## A vanilla plot
absval.iwres.vs.ipred(xpdb)

## A conditioning plot
absval.iwres.vs.ipred(xpdb, by="HCTZ")

## Custom heading and axis labels
absval.iwres.vs.ipred(xpdb, main="My conditioning plot", ylb="|IWRES|", xlb="IPRED")

## Custom colours and symbols, no IDs
absval.iwres.vs.ipred(xpdb, cex=0.6, pch=3, col=1, ids=FALSE)


}
\seealso{
\code{\link{xpose.plot.default}},
\code{\link{xpose.panel.default}}, \code{\link[lattice]{xyplot}},
\code{\link{xpose.prefs-class}}, \code{\link{xpose.data-class}},
\code{\link{runsum}}
}
\author{
E. Niclas Jonsson, Mats Karlsson, Andrew Hooker & Justin Wilkins
}
\keyword{methods}
