% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/absval.iwres.wres.vs.ipred.pred.R
\name{absval.iwres.wres.vs.ipred.pred}
\alias{absval.iwres.wres.vs.ipred.pred}
\title{Plots of the abslute value of the 
individual weighted residuals vs. individual predictions
and the weighted residuals vs the population predictions}
\usage{
absval.iwres.wres.vs.ipred.pred(object, main = "Default", ...)
}
\arguments{
\item{object}{An "xpose.data" object.}

\item{main}{A string giving the plot title or \code{NULL} if none.}

\item{...}{Other arguments passed to \code{\link{xpose.panel.default}}.}
}
\value{
An xpose.multiple.plot object
}
\description{
Plots of the abslute value of the 
individual weighted residuals vs. individual predictions
and the weighted residuals vs the population predictions
}
\seealso{
Other specific functions: \code{\link{absval.cwres.vs.cov.bw}},
  \code{\link{absval.cwres.vs.pred.by.cov}},
  \code{\link{absval.cwres.vs.pred}},
  \code{\link{absval.iwres.cwres.vs.ipred.pred}},
  \code{\link{absval.iwres.vs.cov.bw}},
  \code{\link{absval.iwres.vs.idv}},
  \code{\link{absval_delta_vs_cov_model_comp}},
  \code{\link{autocorr.iwres}}, \code{\link{par_cov_hist}},
  \code{\link{par_cov_qq}}, \code{\link{par_cov_summary}},
  \code{\link{xpose4-package}}
}
