% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.grid.table.R
\name{add.grid.table}
\alias{add.grid.table}
\alias{add.grid.text}
\title{Print tables or text in a grid object}
\usage{
add.grid.table(txt, col.nams = NULL, ystart, xstart = unit(0, "npc"),
  start.pt = 1, vp, vp.num = 1, minrow = 5, cell.padding = 0.5,
  mult.col.padding = 1, col.optimize = TRUE, equal.widths = FALSE,
  space.before.table = 1, center.table = FALSE, use.rect = FALSE,
  fill.type = NULL, fill.col = "grey", cell.lines.lty = 0, ...)
}
\arguments{
\item{txt}{The text or table values to add to the grid object.}

\item{col.nams}{the column names of the table values}

\item{ystart}{The y location to start printing in the grid viewport}

\item{xstart}{The x location to start printing in the grid viewport}

\item{start.pt}{The start point (row) in the table array to start printing}

\item{vp}{The viewport(s) to add the table or text to}

\item{vp.num}{the viewport number in \code{vp} to start printing to}

\item{minrow}{The minimum rows before printing more columns to use in the
table}

\item{cell.padding}{padding between cells in the table}

\item{mult.col.padding}{padding between multiple columns in the table}

\item{col.optimize}{should we column optimize (\code{TRUE}) or row optimize
(\code{FALSE})}

\item{equal.widths}{Should all columns have equal widths}

\item{space.before.table}{Should there be a space before the table}

\item{center.table}{should we center the table in the viewport?}

\item{use.rect}{Should we make rectangles with background color around the
table entries \code{TRUE} or \code{FALSE}}

\item{fill.type}{Which rectangles should be filled.  Allowed values are
\code{"all"}, \code{"top"}, \code{"side"}, \code{"both"} and \code{NULL}.}

\item{fill.col}{The color of the filled rectangles}

\item{cell.lines.lty}{The line-type for the lines between the cells, using
the same values as lty.}

\item{\dots}{Other arguments passed to the various functions.}
}
\value{
A List is returned with the following components \item{ystart }{new
starting point for new text} \item{stop.pt }{null if everything gets
printed} \item{vp.num }{the viewport needed for next text printed}
\item{xpose.table }{A grob object that can be plotted.}
}
\description{
These functions take an array of values and labels or an array of text and
add it to one or many grid viewports in an orderly fashion.
}
\seealso{
\code{\link{runsum}}, \code{\link[grid]{grid.text}}
}
\author{
Andrew Hooker
}
\keyword{methods}
