% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addid.R, R/computePI.R, R/create.rand.R,
%   R/create.strat.rand.R, R/eq.xpose.R, R/get.refrunno.R, R/xpose.stack.R
\name{addid}
\alias{addid}
\alias{computePI}
\alias{create.rand}
\alias{create.strat.rand}
\alias{eq.xpose}
\alias{get.refrunno}
\alias{xpose.stack}
\title{Generic internal functions for Xpose 4}
\usage{
addid(x, y, ids = ids, idsmode = NULL, idsext = 0.05, idscex = 0.7,
  idsdir = "both", gridmode = TRUE)

computePI(x, y, object, limits = object@Prefs@Graph.prefs$PIlimits,
  logy = FALSE, logx = FALSE, onlyfirst = FALSE, inclZeroWRES = FALSE,
  PI.subset = NULL, subscripts)

create.rand(data, object, frac, seed = NULL)

create.strat.rand(data, object, x, y, frac, dilci, seed = NULL)

eq.xpose(x, number = 6, overlap = 0.5)

get.refrunno(database = ".ref.db")

xpose.stack(data, object, select, rep, subset = NULL, ...)
}
\value{
Internal helper functions for the generic Xpose functions.
}
\description{
These are internal functions relating to the Xpose generic functions.
}
\details{
These are internal Xpose functions, for adding ID numbers, computing
prediction intervals, randomization, stacking, and binning. They are not
intended for direct use.
}
\author{
Justin Wilkins and Andrew Hooker
}
\keyword{internal}
