% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootgam.plots.R
\name{bootgam.print}
\alias{bootgam.print}
\title{Print summary information for a bootgam or bootscm}
\usage{
bootgam.print(bootgam.obj = NULL)
}
\arguments{
\item{bootgam.obj}{The bootgam or bootscm object.}
}
\value{
No value returned
}
\description{
This functions prints some summary information for a bootgam performed in
Xpose, or for a bootscm performed in PsN.
}
\examples{

\dontrun{
bootgam.print(current.bootgam)  # Print summary for the current Xpose bootgam object
bootgam.print(current.bootscm)  # Print summary for the current Xpose bootscm object
}

}
\author{
Ron Keizer
}
\keyword{~bootgam}
\keyword{~bootscm}
