% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/categorical.table.R, R/continuous.table.R
\name{categorical.table}
\alias{categorical.table}
\alias{continuous.table}
\alias{continuous.table}
\title{Generic table functions for Xpose 4}
\usage{
categorical.table(object, vars, onlyfirst = TRUE, subset = xsubset(object),
  inclZeroWRES = FALSE, miss = object@Prefs@Miss)

continuous.table(object, vars, onlyfirst = TRUE, subset = xsubset(object),
  inclZeroWRES = FALSE, miss = object@Prefs@Miss)
}
\value{
Internal helper functions for the generic Xpose summary functions.
}
\description{
These are internal table functions relating to the Xpose summary functions.
}
\details{
These are internal Xpose functions for outputting summary tables. They are
not intended for direct use.
}
\author{
Niclas Jonsson, Justin Wilkins and Andrew Hooker
}
\keyword{internal}
