% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.vars.R, R/is.readable.file.R,
%   R/test.xpose.data.R, R/xpose.bin.R
\name{check.vars}
\alias{check.vars}
\alias{is.readable.file}
\alias{test.xpose.data}
\alias{xpose.bin}
\title{Data functions for Xpose 4}
\usage{
check.vars(vars, object, silent = FALSE)

is.readable.file(filename)

test.xpose.data(object)

xpose.bin(data, y, bins = 10)
}
\arguments{
\item{vars}{List of variables to be checked.}

\item{object}{An \code{xpose.data} object.}

\item{silent}{A logical operator specifying whether output should be
displayed.}

\item{filename}{A filename to be checked for readability.}
}
\value{
TRUE, FALSE or NULL.
}
\description{
These functions perform various tasks in managing Xpose data objects.
}
\details{
These are internal Xpose functions, not intended for direct use.
}
\seealso{
\code{\link{xpose.prefs-class}}, \code{\link{xvardef}}
}
\author{
Niclas Jonsson and Andrew Hooker
}
\keyword{internal}
