% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.xpose.plot.classes.R
\name{create.xpose.plot.classes}
\alias{create.xpose.plot.classes}
\alias{print,xpose.multiple.plot-method}
\title{Create xpose.multiple.plot class.}
\usage{
create.xpose.plot.classes()
}
\description{
Creates a class for viewing and plotting xpose plots with multiple plots on
the same page or multiple pages.
}
\author{
Niclas Jonsson and Andrew C. Hooker
}
