% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cwres.vs.idv.R
\name{cwres.vs.idv}
\alias{cwres.vs.idv}
\title{Population conditional weighted residuals (CWRES) plotted against the
independent variable (IDV) for Xpose 4}
\usage{
cwres.vs.idv(object, abline = c(0, 0), smooth = TRUE, ...)
}
\arguments{
\item{object}{An xpose.data object.}

\item{abline}{Vector of arguments to the \code{\link[lattice]{panel.abline}}
function. No abline is drawn if \code{NULL}.}

\item{smooth}{A \code{NULL} value indicates that no superposed line should
be added to the graph. If \code{TRUE} then a smooth of the data will be
superimposed.}

\item{\dots}{Other arguments passed to \code{link{xpose.plot.default}}.}
}
\value{
Returns an xyplot of CWRES vs IDV.
}
\description{
This is a plot of population conditional weighted residuals (CWRES) vs the
independent variable (IDV), a specific function in Xpose 4. It is a wrapper
encapsulating arguments to the \code{xpose.plot.default} function. Most of
the options take their default values from xpose.data object but may be
overridden by supplying them as arguments.
}
\details{
Conditional weighted residuals (CWRES) are plotted against the independent
variable, as specified in \code{object@Prefs@Xvardef$idv}.

Conditional weighted residuals (CWRES) require some extra steps to
calculate. See \code{\link{compute.cwres}} for details.

A wide array of extra options controlling xyplots are available. See
\code{\link{xpose.plot.default}} and \code{\link{xpose.panel.default}} for
details.
}
\examples{

\dontrun{
## We expect to find the required NONMEM run and table files for run
## 5 in the current working directory
xpdb5 <- xpose.data(5)
}

## Here we load the example xpose database 
data(simpraz.xpdb)
xpdb <- simpraz.xpdb

## A vanilla plot
cwres.vs.idv(xpdb)

## A conditioning plot
cwres.vs.idv(xpdb, by="HCTZ")

## Logarithmic Y-axis
cwres.vs.idv(xpdb, logy=TRUE)

## Custom colours and symbols, IDs labels
cwres.vs.idv(xpdb, cex=0.6, pch=3, col=1, ids=TRUE)

}
\seealso{
\code{\link{xpose.plot.default}},
\code{\link{xpose.panel.default}}, \code{\link[lattice]{xyplot}},
\code{\link{xpose.prefs-class}}, \code{\link{compute.cwres}},
\code{\link{xpose.data-class}}
}
\author{
E. Niclas Jonsson, Mats Karlsson, Andrew Hooker & Justin Wilkins
}
\keyword{methods}
