% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cwres.vs.pred.R
\name{cwres.vs.pred}
\alias{cwres.vs.pred}
\title{Population conditional weighted residuals (CWRES) plotted against population
predictions (PRED) for Xpose 4}
\usage{
cwres.vs.pred(object, abline = c(0, 0), smooth = TRUE, ...)
}
\arguments{
\item{object}{An xpose.data object.}

\item{abline}{Vector of arguments to the \code{\link[lattice]{panel.abline}}
function. No abline is drawn if \code{NULL}.}

\item{smooth}{Logical value indicating whether an x-y smooth should be
superimposed.  The default is TRUE.}

\item{\dots}{Other arguments passed to \code{link{xpose.plot.default}}.}
}
\value{
Returns an xyplot of CWRES vs PRED.
}
\description{
This is a plot of population conditional weighted residuals (cwres) vs
population predictions (PRED), a specific function in Xpose 4. It is a
wrapper encapsulating arguments to the \code{xpose.plot.default} function.
Most of the options take their default values from xpose.data object but may
be overridden by supplying them as arguments.
}
\details{
Conditional weighted residuals (CWRES) require some extra steps to
calculate. See \code{\link{compute.cwres}} for details.

A wide array of extra options controlling xyplots are available. See
\code{\link{xpose.plot.default}} and \code{\link{xpose.panel.default}} for
details.
}
\examples{

\dontrun{
## We expect to find the required NONMEM run and table files for run
## 5 in the current working directory
xpdb5 <- xpose.data(5)
}

## Here we load the example xpose database 
data(simpraz.xpdb)
xpdb <- simpraz.xpdb

## A vanilla plot
cwres.vs.pred(xpdb)

## A conditioning plot
cwres.vs.pred(xpdb, by="HCTZ")

## Logarithmic Y-axis
cwres.vs.pred(xpdb, logy=TRUE)

## Custom colours and symbols, IDs
cwres.vs.pred(xpdb, cex=0.6, pch=3, col=1, ids=TRUE)


}
\seealso{
\code{\link{xpose.plot.default}}, \code{\link[lattice]{xyplot}},
\code{\link{xpose.prefs-class}}, \code{\link{compute.cwres}},
\code{\link{xpose.data-class}}
}
\author{
E. Niclas Jonsson, Mats Karlsson, Andrew Hooker & Justin Wilkins
}
\keyword{methods}
