% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cwres.wres.vs.idv.R
\name{cwres.wres.vs.idv}
\alias{cwres.wres.vs.idv}
\alias{cwres.wres.vs.pred}
\title{Weighted residuals (WRES) and conditional WRES (CWRES) plotted against the
independent variable (IDV) or the population predictions (PRED) for Xpose 4}
\usage{
cwres.wres.vs.idv(object, ylb = "Residuals", abline = c(0, 0),
  smooth = TRUE, scales = list(), ...)
}
\arguments{
\item{object}{An xpose.data object.}

\item{ylb}{A string giving the label for the y-axis. \code{NULL} if none.}

\item{abline}{Vector of arguments to the \code{\link[lattice]{panel.abline}}
function. No abline is drawn if \code{NULL}.}

\item{smooth}{A \code{NULL} value indicates that no superposed line should
be added to the graph. If \code{TRUE} then a smooth of the data will be
superimposed.}

\item{scales}{scales is pased to \code{xpose.plot.default}}

\item{\dots}{Other arguments passed to \code{link[lattice]{xyplot}}.}
}
\value{
A compound xyplot.
}
\description{
This function graphically compares WRES and CWRES as plotted against the
independent variable or the popultation predictions.
}
\details{
This function creates plots of WRES and CWRES, presented side-by-side for
comparison.

Conditional weighted residuals (CWRES) require some extra steps to
calculate. See \code{\link{compute.cwres}} for details.

A wide array of extra options controlling xyplots are available. See
\code{\link{xpose.plot.default}} and \code{\link{xpose.panel.default}} for
details.
}
\examples{

\dontrun{
## We expect to find the required NONMEM run, table and data files for run
## 5 in the current working directory 
xpdb5 <- xpose.data(5)
}

## Here we load the example xpose database 
data(simpraz.xpdb)
xpdb <- simpraz.xpdb

cwres.wres.vs.idv(xpdb)

}
\seealso{
\code{\link{xpose.plot.default}},
\code{\link{xpose.panel.default}}, \code{\link[lattice]{xyplot}},
\code{\link{xpose.prefs-class}}, \code{\link{xpose.data-class}},
\code{\link{compute.cwres}}
}
\author{
Niclas Jonsson & Andrew Hooker
}
\keyword{methods}
