% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cwres.wres.vs.idv.R, R/cwres.wres.vs.pred.R
\name{cwres_wres_vs_x}
\alias{cwres_wres_vs_x}
\alias{cwres.wres.vs.idv}
\alias{cwres.wres.vs.pred}
\title{Weighted residuals (WRES) and conditional WRES (CWRES) plotted against the
independent variable (IDV) or the population predictions (PRED) for Xpose 4}
\usage{
cwres.wres.vs.idv(object, ylb = "Residuals", abline = c(0, 0),
  smooth = TRUE, scales = list(), ...)

cwres.wres.vs.pred(object, ylb = "Residuals", abline = c(0, 0),
  smooth = TRUE, scales = list(), ...)
}
\arguments{
\item{object}{An xpose.data object.}

\item{ylb}{A string giving the label for the y-axis. \code{NULL} if none.}

\item{abline}{Vector of arguments to the \code{\link[lattice]{panel.abline}}
function. No abline is drawn if \code{NULL}.}

\item{smooth}{A \code{NULL} value indicates that no superposed line should
be added to the graph. If \code{TRUE} then a smooth of the data will be
superimposed.}

\item{scales}{scales is pased to \code{xpose.plot.default}}

\item{\dots}{Other arguments passed to \code{link[lattice]{xyplot}}.}
}
\value{
A compound xyplot.
}
\description{
These functions graphically compare WRES and CWRES as plotted against the
independent variable or the popultation predictions.
}
\details{
This function creates plots of WRES and CWRES, presented side-by-side for
comparison.

Conditional weighted residuals (CWRES) require some extra steps to
calculate. See \code{\link{compute.cwres}} for details.

A wide array of extra options controlling xyplots are available. See
\code{\link{xpose.plot.default}} and \code{\link{xpose.panel.default}} for
details.
}
\section{Functions}{
\itemize{
\item \code{cwres.wres.vs.idv}: Weighted residuals (WRES) and conditional WRES (CWRES) plotted against the
independent variable (IDV)

\item \code{cwres.wres.vs.pred}: Weighted residuals (WRES) and conditional WRES (CWRES) plotted against the
population predictions (PRED)
}}

\examples{
## Here we load the example xpose database 
xpdb <- simpraz.xpdb

cwres.wres.vs.idv(xpdb)
}
\seealso{
\code{\link{xpose.plot.default}},
\code{\link{xpose.panel.default}}, \code{\link[lattice]{xyplot}},
\code{\link{xpose.prefs-class}}, \code{\link{xpose.data-class}},
\code{\link{compute.cwres}}

Other specific functions: \code{\link{absval.cwres.vs.cov.bw}},
  \code{\link{absval.cwres.vs.pred.by.cov}},
  \code{\link{absval.cwres.vs.pred}},
  \code{\link{absval.iwres.cwres.vs.ipred.pred}},
  \code{\link{absval.iwres.vs.cov.bw}},
  \code{\link{absval.iwres.vs.idv}},
  \code{\link{absval_delta_vs_cov_model_comp}},
  \code{\link{autocorr.iwres}}, \code{\link{cov.splom}},
  \code{\link{ind.plots.cwres.hist}},
  \code{\link{ind.plots.cwres.qq}},
  \code{\link{par_cov_hist}}, \code{\link{par_cov_qq}},
  \code{\link{par_cov_summary}},
  \code{\link{xpose4-package}}
}
\author{
Niclas Jonsson & Andrew Hooker
}
