% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.checkout.R
\name{data.checkout}
\alias{data.checkout}
\title{Check through the source dataset to detect problems}
\usage{
data.checkout(obj = NULL, datafile = ".ask.", hlin = -99,
  dotcol = "black", dotpch = 16, dotcex = 1, idlab = "ID", csv = NULL,
  main = "Default", ...)
}
\arguments{
\item{obj}{NULL or an xpose.data object.}

\item{datafile}{A data file, suitable for import by
\code{\link{read.table}}.}

\item{hlin}{An integer, specifying the line number on which the column
headers appear.}

\item{dotcol}{Colour for the dots in the dotplot. If obj is an xpose data
object then the default is to use the same value as defined for
box-and-whisker plots.}

\item{dotpch}{Plotting character for the dots in the dotplot. If obj is an
xpose data object then the default is to use the same value as defined for
box-and-whisker plots.}

\item{dotcex}{Relative scaling for the dots in the dotplot.  If obj is an
xpose data object then the default is to use the same value as defined for
box-and-whisker plots.}

\item{idlab}{The ID column label in the dataset.  Input as a text string.}

\item{csv}{Is the data file in CSV format (comma separated values)?  If the
value is \code{NULL} then the user is asked at the command line.  If
supplied to the function the value can be \code{TRUE/FALSE}.}

\item{main}{The title to the plot. "default" means that Xpose creates a
title.}

\item{\dots}{Other arguments passed to \code{link[lattice]{dotplot}}.}
}
\value{
A stack of dotplots.
}
\description{
This function graphically "checks out" the dataset to identify errors or
inconsistencies.
}
\details{
This function creates a series of \code{dotplots}, one for each variable in
the dataset, aginst individual ID. Outliers and clusters may easily be
detected in this manner.
}
\examples{

\dontrun{
## We expect to find the required NONMEM run, table and data files for run
## 5 in the current working directory 
xpdb5 <- xpose.data(5)

data.checkout(xpdb5, datafile = "mydata.dta")
data.checkout(datafile = "mydata.dta")
}

}
\seealso{
\code{\link[lattice]{dotplot}}, \code{\link{xpose.prefs-class}},
\code{\link{read.table}}
}
\author{
Niclas Jonsson, Andrew Hooker & Justin Wilkins
}
\keyword{methods}
