% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.names.R
\name{db.names}
\alias{db.names}
\title{Prints the contents of an Xpose data object}
\usage{
db.names(object)
}
\arguments{
\item{object}{An \code{xpose.data} object.}
}
\value{
A detailed summary of the contents of a specified
\code{\link{xpose.data}} object.
}
\description{
These functions print a summary of the specified Xpose object to the R
console.
}
\details{
These functions return a detailed summary of the contents of a specified
\code{\link{xpose.data}} object.
}
\examples{
db.names(simpraz.xpdb)

}
\seealso{
\code{\link{xpose.data}}

Other data functions: \code{\link{add_transformed_columns}},
  \code{\link{change_graphical_parameters}},
  \code{\link{change_misc_parameters}},
  \code{\link{compute.cwres}}, \code{\link{data.checkout}},
  \code{\link{data_extract_or_assign}},
  \code{\link{export.graph.par}},
  \code{\link{export.variable.definitions}},
  \code{\link{import.graph.par}},
  \code{\link{import.variable.definitions}},
  \code{\link{make.sb.data}}, \code{\link{nsim}},
  \code{\link{par_cov_summary}},
  \code{\link{read.TTE.sim.data}},
  \code{\link{read.nm.tables}},
  \code{\link{read_NM_output}},
  \code{\link{read_nm_table}},
  \code{\link{simprazExample}},
  \code{\link{tabulate.parameters}}, \code{\link{xlabel}},
  \code{\link{xpose.data}}, \code{\link{xpose.print}},
  \code{\link{xpose4-package}}, \code{\link{xsubset}}
}
\author{
Niclas Jonsson & Justin Wilkins
}
\concept{data functions}
\keyword{methods}
