% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dv.preds.vs.idv.R
\name{dv.preds.vs.idv}
\alias{dv.preds.vs.idv}
\title{Observations (DV), individual predictions (IPRED) and population predictions
(IPRED) plotted against the independent variable (IDV), for Xpose 4}
\usage{
dv.preds.vs.idv(object, ylb = "Observations/Predictions", layout = c(3, 1),
  smooth = TRUE, scales = list(), ...)
}
\arguments{
\item{object}{An xpose.data object.}

\item{ylb}{A string giving the label for the y-axis. \code{NULL} if none.}

\item{layout}{A list controlling the number of columns and rows in a
compound plot. The default is 2 columns and 1 row.}

\item{smooth}{Logical value indicating whether an x-y smooth should be
superimposed.  The default is TRUE.}

\item{scales}{A list to be used for the \code{scales} argument in
\code{xyplot}.}

\item{\dots}{Other arguments passed to \code{link{xpose.plot.default}}.}
}
\value{
Returns a compound plot comprising plots of observations (DV),
individual predictions (IPRED), and population predictions (PRED) against
the independent variable (IDV).
}
\description{
This is a compound plot consisting of plots of observations (DV), individual
predictions (IPRED), and population predictions (PRED) against the
independent variable (IDV), a specific function in Xpose 4.  It is a wrapper
encapsulating arguments to the \code{xpose.plot.default} function.
}
\details{
A wide array of extra options controlling \code{xyplots} are available. See
\code{\link{xpose.plot.default}} and \code{\link{xpose.panel.default}} for
details.
}
\examples{

\dontrun{
## We expect to find the required NONMEM run and table files for run
## 5 in the current working directory
xpdb5 <- xpose.data(5)
}

## Here we load the example xpose database 
data(simpraz.xpdb)
xpdb <- simpraz.xpdb

## A vanilla plot
dv.preds.vs.idv(xpdb)

## Custom colours and symbols, IDs
dv.preds.vs.idv(xpdb, cex=0.6, pch=8, col=1, ids=TRUE)

}
\seealso{
\code{\link{dv.vs.idv}}, \code{\link{ipred.vs.idv}},
\code{\link{pred.vs.idv}}, \code{\link{xpose.plot.default}},
\code{\link{xpose.panel.default}}, \code{\link[lattice]{xyplot}},
\code{\link{xpose.prefs-class}}, \code{\link{xpose.data-class}}
}
\author{
E. Niclas Jonsson, Mats Karlsson, Andrew Hooker & Justin Wilkins
}
\keyword{methods}
