% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dv.vs.idv.R
\name{dv.vs.idv}
\alias{dv.vs.idv}
\title{Observations (DV) plotted against the independent variable (IDV) for Xpose 4}
\usage{
dv.vs.idv(object, smooth = TRUE, ...)
}
\arguments{
\item{object}{An xpose.data object.}

\item{smooth}{Logical value indicating whether an x-y smooth should be
superimposed.  The default is TRUE.}

\item{\dots}{Other arguments passed to \code{link{xpose.plot.default}}.}
}
\value{
Returns an xyplot of DV vs IDV.
}
\description{
This is a plot of observations (DV) vs the independent variable (IDV), a
specific function in Xpose 4. It is a wrapper encapsulating arguments to the
\code{xpose.plot.default} function. Most of the options take their default
values from xpose.data object but may be overridden by supplying them as
arguments.
}
\details{
A wide array of extra options controlling \code{\link[lattice]{xyplot}} are
available. See \code{\link{xpose.plot.default}} and
\code{\link{xpose.panel.default}} for details.
}
\examples{

\dontrun{
## We expect to find the required NONMEM run and table files for run
## 5 in the current working directory
xpdb5 <- xpose.data(5)
}

## Here we load the example xpose database 
data(simpraz.xpdb)
xpdb <- simpraz.xpdb

## A vanilla plot
dv.vs.idv(xpdb)

## A conditioning plot
dv.vs.idv(xpdb, by="HCTZ")

## Logarithmic Y-axis
dv.vs.idv(xpdb, logy=TRUE)

## Custom colours and symbols, IDs
dv.vs.idv(xpdb, cex=0.6, pch=3, col=1, ids=TRUE)

}
\seealso{
\code{\link{xpose.plot.default}},
\code{\link{xpose.panel.default}}, \code{\link[lattice]{xyplot}},
\code{\link{xpose.prefs-class}}, \code{\link{xpose.data-class}}
}
\author{
E. Niclas Jonsson, Mats Karlsson, Andrew Hooker & Justin Wilkins
}
\keyword{methods}
