% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dv.vs.pred.by.cov.R
\name{dv.vs.pred.by.cov}
\alias{dv.vs.pred.by.cov}
\title{Dependent variable vs population predictions, conditioned on covariates, for
Xpose 4}
\usage{
dv.vs.pred.by.cov(object, abline = c(0, 1), smooth = TRUE,
  main = "Default", ...)
}
\arguments{
\item{object}{An xpose.data object.}

\item{abline}{Vector of arguments to the \code{\link[lattice]{panel.abline}}
function. No abline is drawn if \code{NULL}.}

\item{smooth}{Logical value indicating whether an x-y smooth should be
superimposed.  The default is TRUE.}

\item{main}{The title of the plot.  If \code{"Default"} then a default title
is plotted. Otherwise the value should be a string like \code{"my title"} or
\code{NULL} for no plot title.}

\item{\dots}{Other arguments passed to \code{link{xpose.plot.default}}.}
}
\value{
Returns a stack of xyplots of DV vs PRED, conditioned on covariates.
}
\description{
This is a plot of the dependent variable (DV) vs population predictions
(PRED) conditioned by covariates, a specific function in Xpose 4.  It is a
wrapper encapsulating arguments to the \code{xpose.plot.default} function.
Most of the options take their default values from xpose.data object but may
be overridden by supplying them as arguments.
}
\details{
Each of the covariates in the Xpose data object, as specified in
\code{object@Prefs@Xvardef$Covariates}, is evaluated in turn, creating a
stack of plots.

A wide array of extra options controlling \code{xyplot}s are available. See
\code{\link{xpose.plot.default}} and \code{\link{xpose.panel.default}} for
details.
}
\examples{

\dontrun{
## We expect to find the required NONMEM run and table files for run
## 5 in the current working directory
xpdb5 <- xpose.data(5)

## Here we load the example xpose database 
data(simpraz.xpdb)
xpdb <- simpraz.xpdb

## A vanilla plot
dv.vs.pred.by.cov(xpdb)

## Custom axis labels
dv.vs.pred.by.cov(xpdb, ylb="DV", xlb="PRED")

## Custom colours and symbols, IDs
dv.vs.pred.by.cov(xpdb, cex=0.6, pch=3, col=1, ids=TRUE)
}

}
\seealso{
\code{\link{dv.vs.pred}}, \code{\link{xpose.plot.default}},
\code{\link{xpose.panel.default}}, \code{\link[lattice]{xyplot}},
\code{\link{xpose.prefs-class}}, \code{\link{xpose.data-class}}
}
\author{
E. Niclas Jonsson, Mats Karlsson, Andrew Hooker & Justin Wilkins
}
\keyword{methods}
