% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.graph.par.R
\name{export.graph.par}
\alias{export.graph.par}
\alias{xpose.write}
\title{Exports Xpose graphics settings to a file.}
\usage{
export.graph.par(object)
}
\arguments{
\item{object}{An \code{xpose.data} object.}

\item{file}{The file to contain exported Xpose settings.}
}
\value{
Null.
}
\description{
This function exports graphics settings for a specified Xpose data object to
a file.
}
\details{
This function exports the graphics settings (contents of
object@Prefs@Graph.prefs) for a given \code{xpose.data} object to a file,
typically 'xpose.ini'. It is a wrapper for \code{xpose.write}. Note that the
file format is not the same as is used in
\code{\link{import.variable.definitions}} and
\code{\link{export.variable.definitions}}.
}
\examples{

\dontrun{
## xpdb5 is an Xpose data object
## We expect to find the required NONMEM run and table files for run
## 5 in the current working directory
xpdb5 <- xpose.data(5)

## For a filename prompt
export.graph.par(xpdb5)

## Command-line driven
xpose.write(xpdb5, "c:/XposeSettings/mytheme.ini")
}

}
\seealso{
\code{\link{import.graph.par}}, \code{\link{xpose.prefs-class}}
}
\author{
Niclas Jonsson & Justin Wilkins
}
\keyword{methods}
