% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.variable.definitions.R
\name{export.variable.definitions}
\alias{export.variable.definitions}
\title{Exports Xpose variable definitions to a file from an Xpose data object.}
\usage{
export.variable.definitions(object)
}
\arguments{
\item{object}{An \code{xpose.data} object.}
}
\value{
Null.
}
\description{
This function exports variable definitions for a specified Xpose data object
to a file.
}
\details{
This function exports variable defintions (contents of object@Prefs@Xvardef)
for a given \code{xpose.data} object to a file, typically
'xpose.vardefs.ini'.  Note that file format is not the same as used for
graphics settings. It is a wrapper for the R function \code{\link{dput}}.
}
\examples{

\dontrun{
## xpdb5 is an Xpose data object
## We expect to find the required NONMEM run and table files for run
## 5 in the current working directory
xpdb5 <- xpose.data(5)

export.variable.definitions(xpdb5)
}

}
\seealso{
\code{\link{import.variable.definitions}},
\code{\link{xpose.prefs-class}} \code{\link{dput}}
}
\author{
Niclas Jonsson & Justin Wilkins
}
\keyword{methods}
