% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PI.helper.R
\name{find.right.table}
\alias{find.right.table}
\alias{setup.PPI}
\alias{get.polygon.regions}
\title{Internal functions for the VPC}
\usage{
find.right.table(object, inclZeroWRES, onlyfirst, samp, PI.subset,
  subscripts, PI.bin.table, panel.number, ...)

setup.PPI(PIlimits, PI.mirror, tmp.table, ...)

get.polygon.regions(PPI, PI.mirror, ...)
}
\arguments{
\item{object}{Xpose object}

\item{inclZeroWRES}{Include row sof data with WRES=0}

\item{onlyfirst}{Use only first data for each individual}

\item{samp}{sample number}

\item{PI.subset}{Prediction interval subset}

\item{subscripts}{subscripts}

\item{PI.bin.table}{prediction interval binning table}

\item{panel.number}{panel number}

\item{\dots}{Extra options passed to arguments}

\item{PIlimits}{Prediction interval limits}

\item{PI.mirror}{Prediction interval mirror}

\item{tmp.table}{temporary table}

\item{PPI}{Plot prediction intervals}
}
\value{
Returned to \code{\link{xpose.VPC}}
}
\description{
Internal functions for the VPC
}
\keyword{internal}
