% Xpose 4
% An R-based population pharmacokinetic/
% pharmacodynamic model building aid for NONMEM.
% Copyright (C) 1998-2004 E. Niclas Jonsson and Mats Karlsson.
% Copyright (C) 2005-2008 Andrew C. Hooker, Justin J. Wilkins, 
% Mats O. Karlsson and E. Niclas Jonsson.
% Copyright (C) 2009-2010 Andrew C. Hooker, Mats O. Karlsson and 
% E. Niclas Jonsson.

% This file is a part of Xpose 4.
% Xpose 4 is free software; you can redistribute it and/or
% modify it under the terms of the GNU Lesser General Public License
% as published by the Free Software Foundation, either version 3
% of the License, or (at your option) any later version.

% This program is distributed in the hope that it will be useful,
% but WITHOUT ANY WARRANTY; without even the implied warranty of
% MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
% GNU Lesser General Public License for more details.

% You should have received a copy of the GNU Lesser General Public License
% along with this program.  A copy can be cound in the R installation
% directory under \share\licenses. If not, see http://www.gnu.org/licenses/.

\name{generic}
\alias{addid}
\alias{computePI}
\alias{create.rand}
\alias{create.strat.rand}
\alias{eq.xpose}
\alias{get.refrunno}
\alias{xpose.stack}
%\alias{xpose.bin}
\title{Generic internal functions for Xpose 4}
\description{
  These are internal functions relating to the Xpose generic functions.
}
\details{
  These are internal Xpose functions, for adding ID numbers, computing 
  prediction intervals, randomization, stacking, and binning. They are 
  not intended for direct use.
}
\value{
  Internal helper functions for the generic Xpose functions.
}
\author{Justin Wilkins}
\keyword{methods}
