% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.variable.definitions.R
\name{import.variable.definitions}
\alias{import.variable.definitions}
\title{Imports Xpose variable definitions from a file to an Xpose data object.}
\usage{
import.variable.definitions(object, classic = FALSE)
}
\arguments{
\item{object}{An \code{xpose.data} object.}

\item{classic}{A logical operator specifying whether the function should
assume the classic menu system. This is an internal option and need never be
called from the command line.}
}
\value{
An \code{\link{xpose.data}} object (classic == FALSE) or null
(classic == TRUE).
}
\description{
This function imports variable definitions for a specified Xpose data object
from a file.
}
\details{
This function imports variable defintions (contents of object@Prefs@Xvardef)
for a given \code{xpose.data} object from a file, typically
'xpose.vardefs.ini'.  It returns an \code{xpose.data} object. Note that file
format is not the same as used for graphics settings. It is a wrapper for
the R function \code{\link{dget}}.
}
\examples{

\dontrun{
## xpdb5 is an Xpose data object
## We expect to find the required NONMEM run and table files for run
## 5 in the current working directory
xpdb5 <- xpose.data(5)

xpdb5 <- import.variable.definitions(xpdb5)
}

}
\seealso{
\code{\link{export.variable.definitions}},
\code{\link{xpose.prefs-class}} \code{\link{dget}}
}
\author{
Niclas Jonsson & Justin Wilkins
}
\keyword{methods}
