% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iwres.dist.qq.R
\name{iwres.dist.qq}
\alias{iwres.dist.qq}
\title{Quantile-quantile plot of individual weighted residuals (IWRES), for Xpose 4}
\usage{
iwres.dist.qq(object, ...)
}
\arguments{
\item{object}{An xpose.data object.}

\item{\dots}{Other arguments passed to \code{link{xpose.plot.qq}}.}
}
\value{
Returns a QQ plot of individual weighted residuals (IWRES).
}
\description{
This is a QQ plot of the distribution of individual weighted residuals
(IWRES) in the dataset, a specific function in Xpose 4. It is a wrapper
encapsulating arguments to the \code{xpose.plot.qq} function.
}
\details{
Displays a QQ plot of the individual weighted residuals (IWRES).
}
\examples{

\dontrun{
## We expect to find the required NONMEM run and table files for run
## 5 in the current working directory
xpdb5 <- xpose.data(5)
}

## Here we load the example xpose database 
data(simpraz.xpdb)
xpdb <- simpraz.xpdb

## A vanilla plot
iwres.dist.qq(xpdb)

## A different plotting character
iwres.dist.qq(xpdb, pch=4)


}
\seealso{
\code{\link{xpose.plot.qq}}, \code{\link{xpose.panel.qq}},
\code{\link[lattice]{qqmath}}, \code{\link{xpose.prefs-class}},
\code{\link{xpose.data-class}}
}
\author{
E. Niclas Jonsson, Mats Karlsson, Andrew Hooker & Justin Wilkins
}
\keyword{methods}
