% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.sb.data.R
\name{make.sb.data}
\alias{make.sb.data}
\title{Make stacked bar data set.}
\usage{
make.sb.data(data, idv, dv, nbins = 6, by = NULL, by.nbins = 6, ...)
}
\arguments{
\item{data}{Data set to transform.}

\item{idv}{the idenpendent variable.}

\item{dv}{the dependent variable.}

\item{nbins}{the number of bins.}

\item{by}{Conditioning variable.}

\item{by.nbins}{by.nbins.}

\item{\dots}{additional arguments.}
}
\description{
Function to make stacked bar data set for categorical data plots.
}
\author{
The Xpose team.
}
\keyword{methods}
