% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nsim.R
\name{nsim}
\alias{nsim}
\alias{nsim<-}
\title{Extract or set the value of the Nsim slot.}
\usage{
nsim(object)
}
\arguments{
\item{object}{An "xpose.data" object.}

\item{value}{A value.}
}
\description{
Extract or set the value of the Nsim slot of an "xpose.data" object.
}
\examples{

\dontrun{
## xpdb5 is an Xpose data object
## We expect to find the required NONMEM run and table files for run
## 5 in the current working directory
xpdb5 <- xpose.data(5)

## Report number of simulations
nsim(xpdb5)
}

}
\seealso{
\code{\link{xpose.data-class}}
}
\author{
Niclas Jonsson
}
\keyword{methods}
