% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cov.hist.R, R/parm.hist.R, R/ranpar.hist.R
\name{par_cov_hist}
\alias{par_cov_hist}
\alias{cov.hist}
\alias{parm.hist}
\alias{ranpar.hist}
\title{Plot the parameter or covariate distributions using a histogram}
\usage{
cov.hist(object, onlyfirst = TRUE, main = "Default", ...)

parm.hist(object, onlyfirst = TRUE, main = "Default", ...)

ranpar.hist(object, onlyfirst = TRUE, main = "Default", ...)
}
\arguments{
\item{object}{An xpose.data object.}

\item{onlyfirst}{Logical value indicating if only the first row per
individual is included in the plot.}

\item{main}{The title of the plot.  If \code{"Default"} then a default title
is plotted. Otherwise the value should be a string like \code{"my title"} or
\code{NULL} for no plot title.  For \code{"Default"} the function
\code{\link{xpose.multiple.plot.title}} is used.}

\item{\dots}{Other arguments passed to \code{\link{xpose.plot.histogram}}.}
}
\value{
Delivers a stack of histograms.
}
\description{
These functions plot the parameter or covariate values stored in an Xpose
data object using histograms.
}
\details{
Each of the parameters or covariates in the Xpose data object, as specified
in \code{object@Prefs@Xvardef$parms}, \code{object@Prefs@Xvardef$covariates}
or \code{object@Prefs@Xvardef$ranpar} is evaluated in turn, creating a stack
of histograms.

A wide array of extra options controlling histograms are available. See
\code{\link{xpose.plot.histogram}} for details.
}
\section{Functions}{
\itemize{
\item \code{cov.hist}: Covariate distributions

\item \code{parm.hist}: parameter distributions

\item \code{ranpar.hist}: random parameter distributions
}}

\examples{

## Here we load the example xpose database 
xpdb <- simpraz.xpdb

## Parameter histograms
parm.hist(xpdb)

## Covariate distribution, in green
cov.hist(xpdb, hicol=11, hidcol="DarkGreen", hiborder="White")

## Random parameter histograms
ranpar.hist(xpdb)

}
\seealso{
\code{\link{xpose.plot.histogram}},
\code{\link{xpose.panel.histogram}}, \code{\link[lattice]{histogram}},
\code{\link{xpose.data-class}}, \code{\link{xpose.prefs-class}}

Other specific functions: \code{\link{absval.cwres.vs.cov.bw}},
  \code{\link{absval.cwres.vs.pred.by.cov}},
  \code{\link{absval.cwres.vs.pred}},
  \code{\link{absval.iwres.cwres.vs.ipred.pred}},
  \code{\link{absval.iwres.vs.cov.bw}},
  \code{\link{absval.iwres.vs.idv}},
  \code{\link{absval.iwres.wres.vs.ipred.pred}},
  \code{\link{absval_delta_vs_cov_model_comp}},
  \code{\link{autocorr.iwres}}, \code{\link{par_cov_qq}},
  \code{\link{par_cov_summary}},
  \code{\link{xpose4-package}}
}
\author{
Andrew Hooker & Justin Wilkins
}
\keyword{methods}
