% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parm.vs.parm.R
\name{parm.vs.parm}
\alias{parm.vs.parm}
\title{Plot parameters vs other parameters}
\usage{
parm.vs.parm(object, onlyfirst = TRUE, abline = FALSE, smooth = TRUE,
  type = "p", main = "Default", ...)
}
\arguments{
\item{object}{An xpose.data object.}

\item{onlyfirst}{Logical value indicating whether only the first row per
individual is included in the plot.}

\item{abline}{Allows for a line of identity.}

\item{smooth}{Logical value indicating whether an x-y smooth should be
superimposed.  The default is TRUE.}

\item{type}{The plot type - defaults to points only.}

\item{main}{The title of the plot.  If \code{"Default"} then a default title
is plotted. Otherwise the value should be a string like \code{"my title"} or
\code{NULL} for no plot title.}

\item{\dots}{Other arguments passed to \code{xpose.plot.default}.}
}
\value{
Returns a stack of xyplots and histograms of parameters against
parameters.
}
\description{
This function plots the parameter values stored in an Xpose data object
versus each other in a series of graphs.  The mirror functionality is
available for this function.
}
\details{
Each of the parameters in the Xpose data object, as specified in
\code{object@Prefs@Xvardef$parms}, is plotted against the rest, creating a
stack of plots.

A wide array of extra options controlling \code{xyplots} are available. See
\code{\link{xpose.plot.default}} and \code{\link{xpose.panel.default}} for
details.
}
\examples{

\dontrun{
## We expect to find the required NONMEM run and table files for run
## 5 in the current working directory
xpdb <- xpose.data(5)


parm.vs.parm(xpdb)


parm.vs.parm(xpdb,mirror=3)
}

}
\author{
Andrew Hooker
}
\keyword{methods}
