% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.xpose.multiple.plot.R
\name{show.xpose.multiple.plot}
\alias{show.xpose.multiple.plot}
\title{Show an Xpose multiple plot object.}
\usage{
\method{show}{xpose.multiple.plot}(x, ...)
}
\arguments{
\item{x}{Output object from the function \code{\link{xpose.multiple.plot}}.}

\item{\dots}{Additional options passed to function.}
}
\description{
Show an Xpose multiple plot object, which is the output from the function
\code{\link{xpose.multiple.plot}}.
}
\details{
Show method for a plot class.
}
\seealso{
\code{\link{xpose.multiple.plot}}.
}
\author{
Niclas Jonsson and Andrew C. Hooker
}
