% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simprazExample.R
\name{simprazExample}
\alias{simprazExample}
\title{Function to create files for the simulated Prazosin example in Xpose}
\usage{
simprazExample(overwrite = FALSE)
}
\arguments{
\item{overwrite}{Logical. Should the function overwrite files with the same
names already in the current working directory?}
}
\description{
Creates NONMEM data, model and output files for a model of Prazosin using
simulted data.
}
\details{
Creates files in the current working directory named: run1.ext run1.lst
run1.mod simpraz.dta xptab1
}
\examples{

simprazExample()

}
\author{
Niclas Jonsson and Andrew Hooker
}
\keyword{methods}
