% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wres.dist.hist.R
\name{wres.dist.hist}
\alias{wres.dist.hist}
\title{Histogram of weighted residuals (WRES), for Xpose 4}
\usage{
wres.dist.hist(object, ...)
}
\arguments{
\item{object}{An xpose.data object.}

\item{\dots}{Other arguments passed to \code{\link{xpose.plot.histogram}}.}
}
\value{
Returns a histogram of weighted residuals (WRES).
}
\description{
This is a histogram of the distribution of weighted residuals (WRES) in the
dataset, a specific function in Xpose 4. It is a wrapper encapsulating
arguments to the \code{xpose.plot.histogram} function.
}
\details{
Displays a histogram of the weighted residuals (WRES).
}
\examples{

\dontrun{
## We expect to find the required NONMEM run and table files for run
## 5 in the current working directory
xpdb5 <- xpose.data(5)
}

## Here we load the example xpose database 
data(simpraz.xpdb)
xpdb <- simpraz.xpdb

## A vanilla plot
wres.dist.hist(xpdb)


}
\seealso{
\code{\link{xpose.plot.histogram}},
\code{\link{xpose.panel.histogram}}, \code{\link[lattice]{histogram}},
\code{\link{xpose.prefs-class}}, \code{\link{xpose.data-class}}
}
\author{
E. Niclas Jonsson, Mats Karlsson, Andrew Hooker & Justin Wilkins
}
\keyword{methods}
