% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootgam.plots.R
\name{xp.daic.npar.plot}
\alias{xp.daic.npar.plot}
\title{Distribution of difference in AIC}
\usage{
xp.daic.npar.plot(
  bootscm.obj = NULL,
  main = NULL,
  xlb = "Difference in AIC",
  ylb = "Density",
  ...
)
}
\arguments{
\item{bootscm.obj}{a bootscm object.}

\item{main}{The title of the plot}

\item{xlb}{The x-label of the plot}

\item{ylb}{The y-label of the plot}

\item{...}{Additional parameters passed to \code{panel.xyplot} and \code{xyplot}.}
}
\value{
A lattice plot object.
}
\description{
Distribution of difference in AIC
}
\seealso{
Other bootgam: 
\code{\link{xp.dofv.npar.plot}()},
\code{\link{xp.inc.cond.stab.cov}()},
\code{\link{xp.inc.ind.cond.stab.cov}()},
\code{\link{xp.inc.stab.cov}()},
\code{\link{xp.incl.index.cov.ind}()},
\code{\link{xp.incl.index.cov}()}

Other bootscm: 
\code{\link{bootscm.import}()},
\code{\link{xp.dofv.npar.plot}()},
\code{\link{xp.inc.cond.stab.cov}()},
\code{\link{xp.inc.ind.cond.stab.cov}()},
\code{\link{xp.inc.stab.cov}()},
\code{\link{xp.incl.index.cov.ind}()},
\code{\link{xp.incl.index.cov}()}
}
\concept{bootgam}
\concept{bootscm}
