% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootgam.plots.R
\name{xp.dofv.npar.plot}
\alias{xp.dofv.npar.plot}
\title{Distribution of difference in OFV}
\usage{
xp.dofv.npar.plot(bootscm.obj = NULL, main = NULL,
  xlb = "Difference in OFV", ylb = "Density", ...)
}
\arguments{
\item{bootscm.obj}{a bootscm object.}

\item{main}{The title of the plot}

\item{xlb}{The x-label of the plot}

\item{ylb}{The y-label of the plot}

\item{...}{Additional parameters passed to \code{panel.xyplot} and \code{xyplot}.}
}
\value{
A lattice plot object.
}
\description{
Distribution of difference in OFV
}
\seealso{
Other bootgam: \code{\link{xp.daic.npar.plot}},
  \code{\link{xp.inc.cond.stab.cov}},
  \code{\link{xp.inc.ind.cond.stab.cov}},
  \code{\link{xp.inc.stab.cov}},
  \code{\link{xp.incl.index.cov.ind}},
  \code{\link{xp.incl.index.cov}}

Other bootscm: \code{\link{bootscm.import}},
  \code{\link{xp.daic.npar.plot}},
  \code{\link{xp.inc.cond.stab.cov}},
  \code{\link{xp.inc.ind.cond.stab.cov}},
  \code{\link{xp.inc.stab.cov}},
  \code{\link{xp.incl.index.cov.ind}},
  \code{\link{xp.incl.index.cov}}
}

