% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootgam.plots.R
\name{xp.dofv.plot}
\alias{xp.dofv.plot}
\title{OFV difference (optimism) plot.}
\usage{
xp.dofv.plot(
  bootscm.obj = NULL,
  main = NULL,
  xlb = "Difference in OFV",
  ylb = "Density",
  ...
)
}
\arguments{
\item{bootscm.obj}{The bootgam or bootscm object.}

\item{main}{Plot title.}

\item{xlb}{Label for x-axis.}

\item{ylb}{Label for y-axis.}

\item{\dots}{Additional plotting parameters.}
}
\value{
A lattice plot object is returned.
}
\description{
A plot of the difference in OFV between final bootscm models and the
reference final scm model.
}
\author{
Ron Keizer
}
\keyword{~bootgam}
\keyword{~bootscm}
