% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootgam.plots.R
\name{xp.inc.prob}
\alias{xp.inc.prob}
\title{Inclusion frequency plot}
\usage{
xp.inc.prob(bootgam.obj = NULL, boot.type = NULL, main = NULL,
  col = "#6495ED", xlb = NULL, ylb = "Covariate", ...)
}
\arguments{
\item{bootgam.obj}{The bootgam or bootscm object.}

\item{boot.type}{Either "bootgam" or "bootscm". Default is NULL, which means
the user will be asked to make a choice.}

\item{main}{Plot title}

\item{col}{Color used for the plot.}

\item{xlb}{Label for x-axis.}

\item{ylb}{Label for y-axis.}

\item{\dots}{Additional plotting parameters.}
}
\value{
A lattice plot object will be returned.
}
\description{
Plot the inclusion frequencies of covariates in the final models obtained in
a bootgam or bootscm. Covariates are ordered by inclusion frequency.
}
\author{
Ron Keizer
}
\keyword{~bootgam}
\keyword{~bootscm}
