% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_classes.R
\docType{class}
\name{xpose.data-class}
\alias{xpose.data-class}
\alias{numeric_or_NULL-class}
\alias{data.frame_or_NULL-class}
\title{Class xpose.data}
\description{
The xpose.data class is the fundamental data object in Xpose 4. It contains
the data and preferences used in the creation of the Xpose plots and
analyses.
}
\section{Objects from the Class}{
 Objects are most easily created by the
\code{xpose.data} function, whcih reads the appropriate NONMEM table files
and populates the slots of the object.
}

\seealso{
\code{\link{xpose.data}}, \code{\link{Data}}, \code{\link{SData}}
\code{\link{read.nm.tables}}, \code{\link{xpose.prefs-class}}
}
\author{
Niclas Jonsson and Andrew Hooker
}
\keyword{classes}
