% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xpose.dev.new.R
\name{xpose.dev.new}
\alias{xpose.dev.new}
\title{Create a new graphical device for an Xpose plot.}
\usage{
xpose.dev.new(...)
}
\arguments{
\item{\dots}{Additional arguments to a new graphical device.  see
\code{\link[grDevices]{dev.new}}.}
}
\description{
The function uses the code from dev.new(). This is a function to make
dev.new() back compatible with older versions of R (before 2.8.0).
}
\seealso{
\code{\link[grDevices]{dev.new}}.
}
\author{
Andrew Hooker
}
\keyword{methods}
