% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_classes.R
\docType{class}
\name{xpose.multiple.plot-class}
\alias{xpose.multiple.plot-class}
\alias{list_or_NULL-class}
\alias{logical_or_numeric-class}
\alias{character_or_NULL-class}
\title{Class for creating multiple plots in xpose}
\description{
Class for creating multiple plots in xpose
}
\section{Slots}{

\describe{
\item{\code{plotList}}{A list of lattice plots}

\item{\code{plotTitle}}{The plot title}

\item{\code{prompt}}{Should prompts be used}

\item{\code{new.first.window}}{Create a new first window?}

\item{\code{max.plots.per.page}}{How many plots per page?}

\item{\code{title}}{The title}

\item{\code{mirror}}{Are there mirror plots to create}

\item{\code{bql.layout}}{Should we use bql.layout}
}}

