% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xpose.string.print.R
\name{xpose.string.print}
\alias{xpose.string.print}
\title{Print a pretty string.}
\usage{
xpose.string.print(value, fill = 60, file = "")
}
\arguments{
\item{value}{The text to print.}

\item{fill}{How wide should the text be per row.}

\item{file}{Where to print. \code{""} means to the screen.}
}
\description{
Print a string with a certain number of characters per row.
}
\author{
Niclas Jonsson and Andrew C. Hooker
}
\keyword{methods}
