% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xpose4.R
\name{xpose4}
\alias{xpose4}
\title{Classic menu system for Xpose 4}
\usage{
xpose4()
}
\description{
Classic menu system for Xpose 4
}
\examples{

\dontrun{
xpose4()
}

}
\seealso{
Other classic functions: \code{\link{xpose4-package}}
}
\author{
Andrew Hooker
}
\concept{classic functions}
\keyword{methods}
