% Xpose 4
% An R-based population pharmacokinetic/
% pharmacodynamic model building aid for NONMEM.
% Copyright (C) 1998-2004 E. Niclas Jonsson and Mats Karlsson.
% Copyright (C) 2005-2008 Andrew C. Hooker, Justin J. Wilkins, 
% Mats O. Karlsson and E. Niclas Jonsson.
% Copyright (C) 2009-2010 Andrew C. Hooker, Mats O. Karlsson and 
% E. Niclas Jonsson.

% This file is a part of Xpose 4.
% Xpose 4 is free software; you can redistribute it and/or
% modify it under the terms of the GNU Lesser General Public License
% as published by the Free Software Foundation, either version 3
% of the License, or (at your option) any later version.

% This program is distributed in the hope that it will be useful,
% but WITHOUT ANY WARRANTY; without even the implied warranty of
% MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
% GNU Lesser General Public License for more details.

% You should have received a copy of the GNU Lesser General Public License
% along with this program.  A copy can be cound in the R installation
% directory under \share\licenses. If not, see http://www.gnu.org/licenses/.

\name{xpose4classic-internal}
\alias{bg.conv.crit1}
\alias{bg.conv.crit2}
\alias{bootgam.conv.crit1}
\alias{bootgam.conv.crit2}
\alias{bootgam.menu}
\alias{bootgam.plot.menu}
\alias{bootgam.settings.menu}
\alias{bootscm.menu}
\alias{bootscm.plot.menu}
\alias{change.conv.alg}
\alias{exclude.individuals}
\alias{list.bootgam.settings}
\alias{normalize.median}
\alias{set.max.replicates}
\alias{set.seed.number}
\alias{specify.interval}
\alias{specify.start.check}
\alias{specify.start.model}

\title{Internal xpose4classic Functions}
\description{
  Internal xpose4classic functions
}
\details{
  These are not to be called by the user (or in some cases are just
  waiting for proper documentation to be written :).
}
\keyword{ internal }
