% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xsubset.R
\name{xsubset}
\alias{xsubset}
\alias{xsubset<-}
\title{Extract or set the value of the Subset slot.}
\usage{
xsubset(object)
}
\arguments{
\item{object}{An "xpose.data" object.}

\item{value}{A string with the subset expression.}
}
\value{
A string representing the subset expression.
}
\description{
Extract or set the value of the Subset slot of an "xpose.data" object.
}
\details{
The subset string has the same syntax as the subset argument to, e.g.
\code{panel.xyplot}. Note, however, that the "xpose.data" subset is not used
as an argument to \code{panel.xyplot}. It is intended as the subset argument
to the \code{Data} and \code{SData} functions.
}
\examples{

\dontrun{
## xpdb5 is an Xpose data object
## We expect to find the required NONMEM run and table files for run
## 5 in the current working directory
xpdb5 <- xpose.data(5)

xsubset(xpdb5) <- "DV > 0"
xsubset(xpdb5)
}

}
\seealso{
\code{\link{Data}}, \code{\link{SData}}
}
\author{
Niclas Jonsson
}
\keyword{methods}
