% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/absval.iwres.cwres.vs.ipred.pred.R
\name{absval.iwres.cwres.vs.ipred.pred}
\alias{absval.iwres.cwres.vs.ipred.pred}
\alias{absval.iwres.wres.vs.ipred.pred}
\title{Absolute population weighted residuals vs population predictions, and
absolute individual weighted residuals vs individual predictions, for Xpose
4}
\usage{
absval.iwres.cwres.vs.ipred.pred(object, main = "Default", ...)
}
\arguments{
\item{object}{An xpose.data object.}

\item{main}{The title of the plot.  If \code{"Default"} then a default title
is plotted. Otherwise the value should be a string like \code{"my title"} or
\code{NULL} for no plot title.  For \code{"Default"} the function
\code{\link{xpose.multiple.plot.title}} is used.}

\item{\dots}{Other arguments passed to \code{link{xpose.plot.default}}.}
}
\value{
Returns a compound plot.
}
\description{
This is a matrix plot of absolute population weighted residuals (|CWRES|) vs
population predictions (PRED) and absolute individual weighted residuals
(|IWRES|) vs individual predictions (IPRED), a specific function in Xpose 4.
It is a wrapper encapsulating arguments to the \code{absval.cwres.vs.pred}
and \code{absval.iwres.vs.ipred} functions.
}
\details{
The plots created by the \code{absval.wres.vs.pred} and
\code{absval.iwres.vs.ipred} functions are presented side by side for
comparison.

A wide array of extra options controlling xyplots are available. See
\code{\link{xpose.plot.default}} for details.
}
\examples{

\dontrun{
## We expect to find the required NONMEM run and table files for run
## 5 in the current working directory
xpdb5 <- xpose.data(5)
}

## Here we load the example xpose database 
data(simpraz.xpdb)
xpdb <- simpraz.xpdb

## A vanilla plot
absval.iwres.wres.vs.ipred.pred(xpdb)
absval.iwres.cwres.vs.ipred.pred(xpdb)

## Custom colours and symbols
absval.iwres.cwres.vs.ipred.pred(xpdb, cex=0.6, pch=8, col=1)

}
\seealso{
\code{\link{absval.wres.vs.pred}},
\code{\link{absval.iwres.vs.ipred}}, \code{\link{xpose.plot.default}},
\code{\link{xpose.panel.default}}, \code{\link[lattice]{xyplot}},
\code{\link{xpose.prefs-class}}, \code{\link{xpose.data-class}}

Other specific functions: \code{\link{absval.cwres.vs.cov.bw}},
  \code{\link{absval.cwres.vs.pred.by.cov}},
  \code{\link{absval.cwres.vs.pred}},
  \code{\link{absval.iwres.vs.cov.bw}},
  \code{\link{absval.iwres.vs.idv}},
  \code{\link{absval.iwres.wres.vs.ipred.pred}},
  \code{\link{absval_delta_vs_cov_model_comp}},
  \code{\link{autocorr.iwres}}, \code{\link{par_cov_hist}},
  \code{\link{par_cov_qq}}, \code{\link{par_cov_summary}},
  \code{\link{xpose4-package}}
}
\author{
E. Niclas Jonsson, Mats Karlsson, Andrew Hooker & Justin Wilkins
}
\keyword{methods}
