% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/absval.wres.vs.idv.R
\name{absval.wres.vs.idv}
\alias{absval.wres.vs.idv}
\title{Absolute value of (C)WRES vs. independent variable plot in Xpose4.}
\usage{
absval.wres.vs.idv(object, idv = "idv", wres = "Default", ylb = "Default",
  smooth = TRUE, idsdir = "up", type = "p", ...)
}
\arguments{
\item{object}{An xpose.data object.}

\item{idv}{the independent variable.}

\item{wres}{Which weighted residual to use. \code{"Default"} is the CWRES.}

\item{ylb}{Y-axis label.}

\item{smooth}{Logical value indicating whether an x-y smooth should be
superimposed.  The default is TRUE.}

\item{idsdir}{Direction for displaying point labels. The default is "up",
since we are displaying absolute values.}

\item{type}{Type of plot. The default is points only ("p"), but lines ("l")
and both ("b") are also available.}

\item{\dots}{Other arguments passed to \code{link{xpose.plot.default}}.}
}
\value{
Returns an xyplot of |CWRES| vs idv (often TIME, defined by
\code{\link{xvardef}}).
}
\description{
This is a plot of the absolute value of the CWRES (default, other residuals
as an option) vs independent variable, a specific function in Xpose 4. It is
a wrapper encapsulating arguments to the \code{\link{xpose.plot.default}}
function.  Most of the options take their default values from the xpose.data
object but may be overridden by supplying them as arguments.
}
\details{
A wide array of extra options controlling xyplots are available. See
\code{\link{xpose.plot.default}} for details.
}
\examples{

\dontrun{
## We expect to find the required NONMEM run and table files for run
## 5 in the current working directory
xpdb5 <- xpose.data(5)
}

## Here we load the example xpose database 
data(simpraz.xpdb)
xpdb <- simpraz.xpdb

## A vanilla plot
absval.wres.vs.idv(xpdb)

## A conditioning plot
absval.wres.vs.idv(xpdb, by="HCTZ")

## Custom heading and axis labels
absval.wres.vs.idv(xpdb, main="Hello World", ylb="|CWRES|", xlb="IDV")

## Custom colours and symbols
absval.wres.vs.idv(xpdb, cex=0.6, pch=3, col=1)

## using the NPDEs instead of CWRES
absval.wres.vs.idv(xpdb,wres="NPDE")

## subsets
absval.wres.vs.idv(xpdb,subset="TIME<10")


}
\seealso{
\code{\link{xpose.plot.default}},
\code{\link{xpose.panel.default}}, \code{\link[lattice]{xyplot}},
\code{\link{xpose.prefs-class}}, \code{\link{compute.cwres}},
\code{\link{xpose.data-class}} %% ~~objects to See Also as
\code{\link{help}}, ~~~
}
\author{
Andrew Hooker
}
\keyword{methods}
