% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autocorr.iwres.R
\name{autocorr.iwres}
\alias{autocorr.iwres}
\title{autocorrelation of the individual weighted residuals}
\usage{
autocorr.iwres(object, type = "p", smooth = TRUE, ids = F,
  main = "Default", ...)
}
\arguments{
\item{object}{An "xpose.data" object.}

\item{ids}{A logical value indicating whether text labels should be used as
plotting symbols (the variable used for these symbols indicated by the
\code{idlab} xpose data variable).}

\item{main}{A string giving the plot title or \code{NULL} if none.}

\item{...}{Other arguments passed to \code{\link{xpose.panel.default}}.}
}
\value{
A Lattice object
}
\description{
autocorrelation of the individual weighted residuals
}
\seealso{
Other specific functions: \code{\link{absval.cwres.vs.cov.bw}},
  \code{\link{absval.cwres.vs.pred.by.cov}},
  \code{\link{absval.cwres.vs.pred}},
  \code{\link{absval.iwres.cwres.vs.ipred.pred}},
  \code{\link{absval.iwres.vs.cov.bw}},
  \code{\link{absval.iwres.vs.idv}},
  \code{\link{absval.iwres.wres.vs.ipred.pred}},
  \code{\link{absval_delta_vs_cov_model_comp}},
  \code{\link{par_cov_hist}}, \code{\link{par_cov_qq}},
  \code{\link{par_cov_summary}},
  \code{\link{xpose4-package}}
}
