% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autocorr.wres.R
\name{autocorr.wres}
\alias{autocorr.wres}
\title{Autocorrelation of weighted residuals for Xpose 4}
\usage{
autocorr.wres(object, type = "p", smooth = TRUE, ids = F,
  main = "Default", ...)
}
\arguments{
\item{object}{An xpose.data object.}

\item{type}{1-character string giving the type of plot desired. The
following values are possible, for details, see \code{\link{plot}}: '"p"'
for points, '"l"' for lines, '"o"' for overplotted points and lines, '"b"',
'"c"') for (empty if '"c"') points joined by lines, '"s"' and '"S"' for
stair steps and '"h"' for histogram-like vertical lines.  Finally, '"n"'
does not produce any points or lines.}

\item{smooth}{Logical value indicating whether a smooth should be
superimposed.}

\item{ids}{A logical value indicating whether text labels should be used as
plotting symbols (the variable used for these symbols indicated by the
\code{idlab} xpose data variable).}

\item{main}{The title of the plot.  If \code{"Default"} then a default title
is plotted. Otherwise the value should be a string like \code{"my title"} or
\code{NULL} for no plot title.}

\item{\dots}{Other arguments passed to \code{link{xpose.plot.default}}.}
}
\value{
Returns an aotocorrelation plot for weighted population residuals
(WRES) or individual weighted residuals (IWRES).
}
\description{
This is an autocorrelation plot of weighted residuals.  Most of the options
take their default values from the xpose.data object but may be overridden
by supplying them as arguments.
}
\details{
A wide array of extra options controlling xyplots are available. See
\code{\link{xpose.plot.default}} for details.
}
\examples{

\dontrun{
## We expect to find the required NONMEM run and table files for run
## 5 in the current working directory
xpdb5 <- xpose.data(5)
}

## Here we load the example xpose database 
data(simpraz.xpdb)
xpdb <- simpraz.xpdb

## A vanilla plot
autocorr.wres(xpdb)

## A conditioning plot
autocorr.wres(xpdb, dilution=TRUE)

## Custom heading and axis labels
autocorr.wres(xpdb, main="My conditioning plot", ylb="|CWRES|", xlb="PRED")

## Custom colours and symbols, IDs
autocorr.wres(xpdb, cex=0.6, pch=3, col=1, ids=TRUE)

## A vanilla plot with IWRES
autocorr.iwres(xpdb)

}
\seealso{
\code{\link[lattice]{xyplot}}, \code{\link{xpose.prefs-class}},
\code{\link{xpose.data-class}}
}
\author{
E. Niclas Jonsson, Mats Karlsson, Justin Wilkins & Andrew Hooker
}
\keyword{methods}
