% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cat.pc.R
\name{cat.pc}
\alias{cat.pc}
\title{Categorical (visual) predictive check.}
\usage{
cat.pc(object, dv = xvardef("dv", object), idv = xvardef("idv", object),
  level.to.plot = NULL, subset = NULL, histo = T, median.line = F,
  PI.lines = F, xlb = if (histo) {     paste("Proportion of ", dv) } else { 
     paste(idv) }, ylb = if (histo) {     paste("Percent of Total") } else {  
    paste("Proportion of Total") }, main = xpose.create.title.text(NULL, dv,
  "Predictive check of", object, subset = subset, ...), strip = "Default",
  ...)
}
\arguments{
\item{object}{Xpose data object.}

\item{dv}{The dependent variable (e.g. \code{"DV"} or \code{"CP"}.)}

\item{idv}{The indenpent variable (e.g. \code{"TIME"}.)}

\item{level.to.plot}{The levels to plot.}

\item{subset}{Subset of data.}

\item{histo}{If \code{FALSE} then a VPC is created, given that \code{idv} is
defined.}

\item{median.line}{Make a median line?}

\item{PI.lines}{Make prediction interval lines?}

\item{xlb}{Label for x axis.}

\item{ylb}{label for y axis.}

\item{main}{Main title.}

\item{strip}{Defining how the strips should appear in the conditioning
plots.}

\item{\dots}{Extra arguments passed to the function.}
}
\description{
Categorical (visual) predictive check plots.
}
\examples{

\dontrun{
## read in table files
runno <- 45
xpdb <- xpose.data(runno)

## create proportion (visual) predictive check
cat.pc(xpdb,idv=NULL)
cat.pc(xpdb,idv="DOSE")
cat.pc(xpdb,idv="DOSE",histo=F)
cat.pc(xpdb,idv="TIME",histo=T,level.to.plot=1)
}

}
\author{
Andrew C. Hooker
}
\keyword{methods}
