% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change.parm.R
\name{change.parm}
\alias{change.parm}
\title{Change parameter scope.}
\usage{
change.parm(object, listall = TRUE, classic = FALSE)
}
\arguments{
\item{object}{The xpose data opject.}

\item{listall}{whether we should list all the current parameters.}

\item{classic}{true if used in the classic menu system (for internal use).}
}
\value{
If classic then return nothing.  Otherwise return the new data
object.
}
\description{
Function to change the parameter scope.
}
\author{
Andrew C. Hooker
}
\keyword{methods}
