% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.mirror.R
\name{create.mirror}
\alias{create.mirror}
\title{Function to create mirror plots from the generic Xpose plotting commands}
\usage{
create.mirror(fun, arg.list, mirror, plotTitle, fix.y.limits = TRUE,
  fix.x.limits = TRUE, ...)
}
\arguments{
\item{fun}{The function name that we will call multiple times}

\item{arg.list}{The arguments to that function}

\item{mirror}{The type of mirror plots desired (1 or 3 mirror plots can be
created)}

\item{plotTitle}{The title for the plots}

\item{fix.y.limits}{Should we fix all the y axes to be the same?}

\item{fix.x.limits}{Should we fix all the x axes to be the same?}

\item{\dots}{additional arguments passed to the function.}
}
\value{
a list of plots, or NULL.
}
\description{
This function takes the generic plotting fucntions from Xpose 4 and calls
them multiple times with the current arguments to the functions, changing
the arguments as needed for mirror plotting.
}
\details{
mostly and internal function for Xpose
}
\seealso{
\code{\link{xpose.plot.default}},
\code{\link{xpose.plot.histogram}}, \code{\link{xpose.plot.qq}},
\code{\link{xpose.plot.splom}}
}
\author{
Andrew Hooker
}
\keyword{methods}
