% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createXposeClasses.R
\name{createXposeClasses}
\alias{createXposeClasses}
\title{This function creates the Xpose data classes ("xpose.data" and
"xpose.prefs")}
\usage{
createXposeClasses(nm7 = F)
}
\arguments{
\item{nm7}{\code{FALSE} if not using NONMEM 7.}
}
\description{
This function defines and sets the Xpose data classes.
}
\note{
All the default settings are defined in this function.
}
\seealso{
\code{\link{xpose.data-class}},\code{\link{xpose.prefs-class}}
}
\author{
Niclas Jonsson and Andrew C. Hooker
}
\keyword{methods}
