% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cwres.vs.pred.bw.R
\name{cwres.vs.pred.bw}
\alias{cwres.vs.pred.bw}
\title{Box-and-whisker plot of conditional weighted residuals vs population
predictions for Xpose 4}
\usage{
cwres.vs.pred.bw(object, ...)
}
\arguments{
\item{object}{An xpose.data object.}

\item{\dots}{Other arguments passed to \code{link{xpose.plot.bw}}.}
}
\value{
Returns a box-and-whisker plot of CWRES vs PRED.
}
\description{
This creates a box and whisker plot of conditional weighted residuals
(CWRES) vs population predictions (PRED), and is a specific function in
Xpose 4.  It is a wrapper encapsulating arguments to the
\code{xpose.plot.bw} function. Most of the options take their default values
from xpose.data object but may be overridden by supplying them as arguments.
}
\details{
This creates a box and whisker plot of conditional weighted residuals
(CWRES) vs population predictions (PRED), and is a specific function in
Xpose 4.  It is a wrapper encapsulating arguments to the
\code{xpose.plot.bw} function. Most of the options take their default values
from xpose.data object but may be overridden by supplying them as arguments.

Conditional weighted residuals (CWRES) require some extra steps to
calculate. See \code{\link{compute.cwres}} for details.

A wide array of extra options controlling bwplots are available. See
\code{\link{xpose.plot.bw}} and \code{\link{xpose.panel.bw}} for details.
}
\examples{

\dontrun{
## We expect to find the required NONMEM run and table files for run
## 5 in the current working directory
xpdb5 <- xpose.data(5)
}

## Here we load the example xpose database 
data(simpraz.xpdb)
xpdb <- simpraz.xpdb

## A vanilla plot
cwres.vs.pred.bw(xpdb)


}
\seealso{
\code{\link{xpose.plot.bw}}, \code{\link{xpose.panel.bw}},
\code{\link[lattice]{bwplot}}, \code{\link{xpose.prefs-class}},
\code{\link{compute.cwres}}, \code{\link{xpose.data-class}}
}
\author{
E. Niclas Jonsson, Mats Karlsson, Andrew Hooker & Justin Wilkins
}
\keyword{methods}
