% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.names.R
\name{db.names}
\alias{db.names}
\alias{xpose.print}
\title{Prints the contents of an Xpose data object}
\usage{
db.names(object)
}
\arguments{
\item{object}{An \code{xpose.data} object.}

\item{long}{A logical operator specifying whether the function should print
an abridged (long = FALSE) or full list of its contents.}
}
\value{
A detailed summary of the contents of a specified
\code{\link{xpose.data}} object.
}
\description{
These functions print a summary of the specified Xpose object to the R
console.
}
\details{
These functions return a detailed summary of the contents of a specified
\code{\link{xpose.data}} object.
}
\examples{

\dontrun{
## xpdb5 is an Xpose data object
## We expect to find the required NONMEM run and table files for run
## 5 in the current working directory
xpdb5 <- xpose.data(5)

xpose.print(xpdb5)
}

}
\seealso{
\code{\link{xpose.data}}
}
\author{
Niclas Jonsson & Justin Wilkins
}
\keyword{methods}
