% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.sb.data.R
\name{make.sb.data}
\alias{make.sb.data}
\title{Make stacked bar data set.}
\usage{
make.sb.data(data, idv, dv, nbins = 6, by = NULL, by.nbins = 6, ...)
}
\arguments{
\item{data}{Data set to transform.}

\item{idv}{the independent variable.}

\item{dv}{the dependent variable.}

\item{nbins}{the number of bins.}

\item{by}{Conditioning variable.}

\item{by.nbins}{by.nbins.}

\item{\dots}{additional arguments.}
}
\description{
Function to make stacked bar data set for categorical data plots.
}
\seealso{
Other data functions: 
\code{\link{add_transformed_columns}},
\code{\link{change_graphical_parameters}},
\code{\link{change_misc_parameters}},
\code{\link{compute.cwres}()},
\code{\link{data.checkout}()},
\code{\link{data_extract_or_assign}},
\code{\link{db.names}()},
\code{\link{export.graph.par}()},
\code{\link{export.variable.definitions}()},
\code{\link{import.graph.par}()},
\code{\link{import.variable.definitions}()},
\code{\link{nsim}()},
\code{\link{par_cov_summary}},
\code{\link{read.TTE.sim.data}()},
\code{\link{read.nm.tables}()},
\code{\link{read_NM_output}},
\code{\link{read_nm_table}()},
\code{\link{simprazExample}()},
\code{\link{tabulate.parameters}()},
\code{\link{xlabel}()},
\code{\link{xpose.data}},
\code{\link{xpose.print}()},
\code{\link{xpose4-package}},
\code{\link{xsubset}()}
}
\author{
The Xpose team.
}
\concept{data functions}
\keyword{methods}
