% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.TTE.sim.data.R
\name{read.TTE.sim.data}
\alias{read.TTE.sim.data}
\title{Read (repeated) time-to-event simulation data files.}
\usage{
read.TTE.sim.data(sim.file, subset = NULL, headers = c("REP", "ID", "DV",
  "TIME", "FLAG2", "DOSE"), xpose.table.file = FALSE, ...)
}
\arguments{
\item{sim.file}{Name of the simulated file.}

\item{subset}{subset to extract.}

\item{headers}{headers in file.}

\item{xpose.table.file}{xpose table files.}

\item{\dots}{Extra arguments passed to function.}
}
\description{
Read (repeated) time-to-event simulation data files.
}
\author{
Andrew C. Hooker
}
\keyword{methods}
