% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.TTE.sim.data.R
\name{read.TTE.sim.data}
\alias{read.TTE.sim.data}
\title{Read (repeated) time-to-event simulation data files.}
\usage{
read.TTE.sim.data(sim.file, subset = NULL, headers = c("REP", "ID", "DV",
  "TIME", "FLAG2", "DOSE"), xpose.table.file = FALSE, ...)
}
\arguments{
\item{sim.file}{Name of the simulated file.}

\item{subset}{subset to extract.}

\item{headers}{headers in file.}

\item{xpose.table.file}{xpose table files.}

\item{\dots}{Extra arguments passed to function.}
}
\description{
Read (repeated) time-to-event simulation data files.
}
\seealso{
Other data functions: \code{\link{add_transformed_columns}},
  \code{\link{change_graphical_parameters}},
  \code{\link{change_misc_parameters}},
  \code{\link{compute.cwres}}, \code{\link{data.checkout}},
  \code{\link{data_extract_or_assign}},
  \code{\link{db.names}}, \code{\link{export.graph.par}},
  \code{\link{export.variable.definitions}},
  \code{\link{import.graph.par}},
  \code{\link{import.variable.definitions}},
  \code{\link{make.sb.data}}, \code{\link{nsim}},
  \code{\link{par_cov_summary}},
  \code{\link{read.nm.tables}},
  \code{\link{read_NM_output}},
  \code{\link{read_nm_table}},
  \code{\link{simprazExample}},
  \code{\link{tabulate.parameters}}, \code{\link{xlabel}},
  \code{\link{xpose.data}}, \code{\link{xpose.print}},
  \code{\link{xpose4-package}}, \code{\link{xsubset}}
}
\author{
Andrew C. Hooker
}
