% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runsum.R
\name{runsum}
\alias{runsum}
\title{Print run summary in Xpose 4}
\usage{
runsum(object, dir = "", modfile = paste(dir, "run", object@Runno, ".mod",
  sep = ""), listfile = paste(dir, "run", object@Runno, ".lst", sep = ""),
  main = NULL, subset = xsubset(object), show.plots = TRUE,
  txt.cex = 0.7, txt.font = 1, show.ids = FALSE, param.table = TRUE,
  txt.columns = 2, force.wres = FALSE, ...)
}
\arguments{
\item{object}{An xpose.data object.}

\item{dir}{The directory to look for the model and output file of a NONMEM
run.}

\item{modfile}{The name of the NONMEM control stream associated with the
current run.}

\item{listfile}{The name of the NONMEM output file associated with the
current run.}

\item{main}{A string giving the main heading. \code{NULL} if none.}

\item{subset}{A string giving the subset expression to be applied to the
data before plotting. See \code{\link{xsubset}}.}

\item{show.plots}{Logical indicating if GOF plots should be shown in the run
summary.}

\item{txt.cex}{Number indicating the size of the txt in the run summary.}

\item{txt.font}{Font of the text in the run summary.}

\item{show.ids}{Logical indicating if IDs should be plotted in the plots for
the run summary.}

\item{param.table}{Logical indicating if the parameter table should be shown
in the run summary.}

\item{txt.columns}{The number of text columns in the run summary.}

\item{force.wres}{Plot the WRES even if other residuals are available.}

\item{\dots}{Other arguments passed to the various functions.}
}
\value{
A compound plot containing an Xpose run summary is created.
}
\description{
Function to build Xpose run summaries.
}
\examples{

\dontrun{
## We expect to find the required NONMEM run and table files for run
## 5 in the current working directory
xpdb5 <- xpose.data(5)
}

## Here we load the example xpose database 
data(simpraz.xpdb)
xpdb <- simpraz.xpdb

runsum(xpdb)


}
\author{
Niclas Jonsson and Andrew Hooker
}
\keyword{methods}
