% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prazosin_xpdb.R
\docType{data}
\name{simpraz.xpdb}
\alias{simpraz.xpdb}
\title{Simulated prazosin Xpose database.}
\format{an xpose.data object}
\usage{
simpraz.xpdb
}
\description{
Xpose database from the NONMEM output of a model for prazosin using
simulated data (and NONMEM 7.3).
}
\details{
The database can be used to test functions in Xpose 4.  This database is
slightly different than the database that is created when reading in the
files created by \code{\link{simprazExample}} using
\code{\link{xpose.data}}.
}
\examples{

xpose.print(simpraz.xpdb)
Data(simpraz.xpdb)
str(simpraz.xpdb)

}
\seealso{
\code{\link{simprazExample}}
}
\keyword{datasets}
