% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabulate.parameters.R
\name{tabulate.parameters}
\alias{tabulate.parameters}
\title{Tabulate the population parameter estimates}
\usage{
tabulate.parameters(object, prompt = FALSE, outfile = NULL, dir = "")
}
\arguments{
\item{object}{An xpose.data object.}

\item{prompt}{Ask before printing.}

\item{outfile}{file to output to (NULL means screen).}

\item{dir}{Wich directory is the NONMEM output file located. \code{""} means
the current working directory \code{getwd()}.}
}
\value{
A table summarizing the parameters and their precision.
}
\description{
This function provides a summary of the model's parameter estimates and
precision.
}
\examples{

\dontrun{ 
## We expect to find the required NONMEM run and table files for run
## 5 in the current working directory
xpdb5 <- xpose.data(5)

tabulate.parameters(xpdb)
}
}
\author{
Niclas Jonsson, Andrew Hooker & Justin Wilkins
}
\keyword{methods}
