% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xlabel.R
\name{xlabel}
\alias{xlabel}
\alias{xlabel<-}
\title{Extract and set labels for Xpose data items.}
\usage{
xlabel(x, object)

xlabel(object) <- value
}
\arguments{
\item{x}{Name of the variable to assign a label to.}

\item{object}{An \code{xpose.data} object.}

\item{value}{A two element vector of which the first element is the name of
the variable and the second the label}
}
\value{
The label of the specified column.
}
\description{
This function extracts and sets label definitions in Xpose data objects.
}
\details{
\code{x} should be a string exactly matching the name of a column in the
data.frame in the Data slot of an xpose.data object. The name of columns
defined through xpose variable definitions (see \code{\link{xpose.data}})
can be extracted using the \code{xvardef} function and to be used in the
\code{xlabel} function, e.g. \code{xlabel(xvardef("dv",object),object)},
which would give the label for the \code{dv} variable.
}
\section{Functions}{
\itemize{
\item \code{xlabel<-}: sets label definitions in Xpose data objects.  assigned value should be a two-element vector 
of which the first element is the name of
the variable and the second the label
}}

\examples{

\dontrun{
## xpdb5 is an Xpose data object
## We expect to find the required NONMEM run and table files for run
## 5 in the current working directory
xpdb5 <- xpose.data(5)

## Display label for dependent variable in the Xpose data object
xlabel(xvardef("dv", object), object)

## Set label for dependent variable
xlabel(xpdb5) <- c(xvardef("dv", object), "Concentration (mg/L)")
}

}
\seealso{
\code{\link{xpose.prefs-class}}, \code{\link{xvardef}}
}
\author{
Niclas Jonsson
}
\keyword{methods}
