% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootgam.plots.R
\name{xp.distr.mod.size}
\alias{xp.distr.mod.size}
\title{Plot of model size distribution for a bootgam or bootscm}
\usage{
xp.distr.mod.size(bootgam.obj = NULL, boot.type = NULL, main = NULL,
  bw = 0.5, xlb = NULL, ...)
}
\arguments{
\item{bootgam.obj}{The bootgam or bootscm object.}

\item{boot.type}{Either "bootgam" or "bootscm". Default is NULL, which means
the user will be asked to make a choice.}

\item{main}{Plot title.}

\item{bw}{The smoothing bandwidth to be used for the kernel.}

\item{xlb}{The x-axis label.}

\item{\dots}{Additional plotting parameter may be passed to this function.}
}
\value{
A lattice plot object will be returned.
}
\description{
This function creates a kernel smoothed plot of the number of covariates
included in the final model in each gam/scm in the bootgam/bootscm
procedure.
}
\author{
Ron Keizer
}
\keyword{~bootgam}
\keyword{~bootscm}
