% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootgam.plots.R
\name{xp.inc.prob.comb.2}
\alias{xp.inc.prob.comb.2}
\title{Inclusion frequency plot for combination of covariates.}
\usage{
xp.inc.prob.comb.2(bootgam.obj = NULL, boot.type = NULL, main = NULL,
  col = "#6495ED", xlb = NULL, ylb = "Covariate combination", ...)
}
\arguments{
\item{bootgam.obj}{The bootgam or bootscm object.}

\item{boot.type}{Either "bootgam" or "bootscm". Default is NULL, which means
the user will be asked to make a choice.}

\item{main}{Plot title}

\item{col}{Color used for plot.}

\item{xlb}{Label for x-axis.}

\item{ylb}{Label for y-axis.}

\item{\dots}{Additional plotting parameters.}
}
\value{
A lattice plot object will be returned.
}
\description{
Plot the inclusion frequency of the most common 2-covariate combinations.
}
\author{
Ron Keizer
}
\keyword{~bootgam}
\keyword{~bootscm}
