% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootgam.plots.R
\name{xp.incl.index.cov.comp}
\alias{xp.incl.index.cov.comp}
\title{Inclusion index individuals, compare between covariates.}
\usage{
xp.incl.index.cov.comp(
  bootgam.obj = NULL,
  boot.type = NULL,
  main = NULL,
  xlb = "Individual inclusion index",
  ylb = "ID",
  ...
)
}
\arguments{
\item{bootgam.obj}{A bootgam or bootscm object.}

\item{boot.type}{Either "bootgam" or "bootscm". Default is NULL, which means
the user will be asked to make a choice.}

\item{main}{The title of the plot.}

\item{xlb}{The label for the x-axis.}

\item{ylb}{The label for the y-axis.}

\item{\dots}{Additional plotting parameters.}
}
\value{
A lattice plot object is returned.
}
\description{
A plot showing the range of inclusion indices for individuals for all
covariates. This plot can be used to evaluate whether there were covariates
which were more influenced by the constituency of the bootstrapped dataset
than others.
}
\author{
Ron Keizer
}
\keyword{~bootgam}
\keyword{~bootscm}
