% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xpose.ask.for.filename.R
\name{xpose.ask.for.filename}
\alias{xpose.ask.for.filename}
\alias{xpose.ask.for.mod}
\alias{xpose.ask.for.lst}
\title{Function to ask the user for the name of a file}
\usage{
xpose.ask.for.filename(object, listfile = paste("run", object@Runno, ".lst",
  sep = ""), modfile = paste("run", object@Runno, ".mod", sep = ""), ...)
}
\arguments{
\item{object}{An \code{\link{xpose.data}} object.}

\item{listfile}{A NONMEM output file}

\item{modfile}{A NONMEM model file}

\item{\dots}{Additional arguments passed to the function}
}
\value{
The name of the file if it exists, otherwise nothing is returned.
}
\description{
Asks the user for the name of a file.
}
\details{
Function checks if the file exists, if it does then the filename is returned
from the function.
}
\author{
Niclas Jonsson, Justin Wilkins, Mats Karlsson and Andrew Hooker
}
\keyword{methods}
