% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xpose.create.title.R
\name{xpose.create.title}
\alias{xpose.create.title}
\alias{xpose.create.label}
\title{Functions to create labels for plots}
\usage{
xpose.create.title(
  x,
  y,
  object,
  subset = NULL,
  funx = NULL,
  funy = NULL,
  no.runno = FALSE,
  ...
)

xpose.create.label(
  x,
  object,
  fun,
  logx,
  autocorr.x = FALSE,
  autocorr.y = FALSE,
  ...
)
}
\arguments{
\item{x}{Column name for x-variable}

\item{y}{Column name for y variable}

\item{object}{Xpose data object}

\item{subset}{Subset used for plot}

\item{funx}{Function applied to x data}

\item{funy}{Function applied to y data}

\item{no.runno}{should we include a run number in the label}

\item{\dots}{additional arguments passed to the function.}

\item{fun}{Function applied to data}

\item{logx}{is the data log transformed?}

\item{autocorr.x}{Are we looking at an autocorrelation plot?}

\item{autocorr.y}{Are we looking at an autocorrelation plot?}

\item{plot.text}{Text of plot title}

\item{main}{If \code{"Default"} then \code{plot.text} is the default title
with the run number and subset expression added to the title. If the value
is anything other than \code{"Default"} then that value is used as the plot
title.}
}
\value{
Plot titles and labels.
}
\description{
Functions to create labels for plots
}
\section{Functions}{
\itemize{
\item \code{xpose.create.label}: Create label values
}}

\author{
Andrew Hooker
}
\keyword{internal}
