% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xpose.create.title.R
\name{xpose.create.title.text}
\alias{xpose.create.title.text}
\title{Create Xpose title text for plots.}
\usage{
xpose.create.title.text(x, y, text, object, subset, text2 = NULL, ...)
}
\arguments{
\item{x}{The x-axis variable name.}

\item{y}{The y-axis variable name.}

\item{text}{Initial text in title.}

\item{object}{Xpose data object \code{\link{xpose.data}}.}

\item{subset}{Subset definition.}

\item{text2}{Text at the end of the title.}

\item{\dots}{Additional options passed to function.}
}
\description{
Create Xpose title text for plots.
}
\author{
Andrew C. Hooker
}
\keyword{internal}
