% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xpose.license.citation.R
\name{xpose.license.citation}
\alias{xpose.license.citation}
\title{Displays the Xpose license and citation information}
\usage{
xpose.license.citation()
}
\value{
The EULA.
}
\description{
This function displays a copy of Xpose's end user license agreement (EULA).
}
\examples{

xpose.license.citation()

}
\author{
Andrew Hooker
}
\keyword{methods}
