% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_classes.R
\docType{class}
\name{xpose.prefs-class}
\alias{xpose.prefs-class}
\alias{character_or_numeric-class}
\title{Class "xpose.prefs"}
\description{
An object of the "xpose.prefs" class holds information about all the
variable and graphical preferences for a particular "xpose.data" object.
}
\section{Objects from the Class}{
 Objects can be created by calls of the form
\code{new("xpose.prefs",...)} but this is usually not necessary since the
"xpose.prefs" object is created at the same time as the "xpose.data" object.
}

\seealso{
\code{\link{xvardef}}, \code{\link{xlabel}}, \code{\link{xsubset}},
\code{\link{Data}}, \code{\link{SData}}, \code{\link{xpose.data}},
\code{\link{read.nm.tables}}, \code{\link{xpose.data-class}},
\code{\link{xpose.gam}}
}
\author{
Niclas Jonsson & Andrew Hooker
}
\keyword{classes}
