% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xpose.print.R
\name{xpose.print}
\alias{xpose.print}
\title{Summariuze an xpose database}
\usage{
xpose.print(object, long = TRUE)
}
\arguments{
\item{object}{An xpose data object}

\item{long}{long format or not.}
}
\value{
""
}
\description{
Summariuze an xpose database
}
\examples{
xpose.print(simpraz.xpdb)
}
\seealso{
Other data functions: \code{\link{add_transformed_columns}},
  \code{\link{change_graphical_parameters}},
  \code{\link{change_misc_parameters}},
  \code{\link{compute.cwres}}, \code{\link{data.checkout}},
  \code{\link{data_extract_or_assign}},
  \code{\link{db.names}}, \code{\link{export.graph.par}},
  \code{\link{export.variable.definitions}},
  \code{\link{import.graph.par}},
  \code{\link{import.variable.definitions}},
  \code{\link{make.sb.data}}, \code{\link{nsim}},
  \code{\link{par_cov_summary}},
  \code{\link{read.TTE.sim.data}},
  \code{\link{read.nm.tables}},
  \code{\link{read_NM_output}},
  \code{\link{read_nm_table}},
  \code{\link{simprazExample}},
  \code{\link{tabulate.parameters}}, \code{\link{xlabel}},
  \code{\link{xpose.data}}, \code{\link{xpose4-package}},
  \code{\link{xsubset}}
}
