% Xpose 4
% An R-based population pharmacokinetic/
% pharmacodynamic model building aid for NONMEM.
% Copyright (C) 1998-2004 E. Niclas Jonsson and Mats Karlsson.
% Copyright (C) 2005-2008 Andrew C. Hooker, Justin J. Wilkins, 
% Mats O. Karlsson and E. Niclas Jonsson.
% Copyright (C) 2009-2010 Andrew C. Hooker, Mats O. Karlsson and 
% E. Niclas Jonsson.

% This file is a part of Xpose 4.
% Xpose 4 is free software; you can redistribute it and/or
% modify it under the terms of the GNU Lesser General Public License
% as published by the Free Software Foundation, either version 3
% of the License, or (at your option) any later version.

% This program is distributed in the hope that it will be useful,
% but WITHOUT ANY WARRANTY; without even the implied warranty of
% MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
% GNU Lesser General Public License for more details.

% You should have received a copy of the GNU Lesser General Public License
% along with this program.  A copy can be cound in the R installation
% directory under \share\licenses. If not, see http://www.gnu.org/licenses/.

\name{xpose4classic-package}
\alias{xpose4classic-package}
\title{Xpose classic menu system}
\docType{package}
\description{
  Functions used to control the classic menu system in Xpose.
}
\details{ 
Xpose has  a text based menu interface to make it simple for the user to invoke the Xpose specific functions. This interface is called Xpose Classic. Given the limitations a text based interface imposes, Xpose Classic is not very flexible but may  be useful for quick assessment of a model and for learning to use Xpose.

\itemize{
\item \code{\link{add.modify.db.items.menu}}
\item \code{\link{another.menu}}
\item \code{\link{change.dispersion}}
\item \code{\link{change.graphical.par}}
\item \code{\link{change.medianNorm}}
\item \code{\link{change.onlyfirst}}
\item \code{\link{change.steppit}}
\item \code{\link{change.use.ids}}
\item \code{\link{change.xp.obj}}
\item \code{\link{covariate.model.menu}}
\item \code{\link{cwres.menu}}
\item \code{\link{data.checkout.menu}}
\item \code{\link{documentation.menu}}
\item \code{\link{gam.change.weights}}
\item \code{\link{gam.menu}}
\item \code{\link{gam.settings.menu}}
\item \code{\link{gof.menu}}
\item \code{\link{list.gam.settings}}
\item \code{\link{main.menu}}
\item \code{\link{manage.db}}
\item \code{\link{model.comparison.covariates.menu}}
\item \code{\link{model.comparison.menu}}
\item \code{\link{parameters.menu}}
\item \code{\link{preferences.menu}}
\item \code{\link{residual.diagnostics.menu}}
\item \code{\link{runsum.print}}
\item \code{\link{struct.gof.menu}}
\item \code{\link{structural.diagnostics.menu}}
\item \code{\link{table.file.read.settings.menu}}
\item \code{\link{vpc.npc.menu}}
\item \code{\link{xpose4}}
}
}

\author{E. Niclas Jonsson, Mats Karlsson, Justin Wilkins and Andrew Hooker}
\keyword{methods}
\keyword{package}




