% Xpose 4
% An R-based population pharmacokinetic/
% pharmacodynamic model building aid for NONMEM.
% Copyright (C) 1998-2004 E. Niclas Jonsson and Mats Karlsson.
% Copyright (C) 2005-2008 Andrew C. Hooker, Justin J. Wilkins, 
% Mats O. Karlsson and E. Niclas Jonsson.
% Copyright (C) 2009-2010 Andrew C. Hooker, Mats O. Karlsson and 
% E. Niclas Jonsson.

% This file is a part of Xpose 4.
% Xpose 4 is free software; you can redistribute it and/or
% modify it under the terms of the GNU Lesser General Public License
% as published by the Free Software Foundation, either version 3
% of the License, or (at your option) any later version.

% This program is distributed in the hope that it will be useful,
% but WITHOUT ANY WARRANTY; without even the implied warranty of
% MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
% GNU Lesser General Public License for more details.

% You should have received a copy of the GNU Lesser General Public License
% along with this program.  A copy can be cound in the R installation
% directory under \share\licenses. If not, see http://www.gnu.org/licenses/.

\name{xpose4data-package}
\alias{xpose4data-package}
\title{Xpose database creation and manipulation}
\docType{package}
\description{
  Functions used to read in NONMEM table files to an Xpose database
  and to create and manipulate
  an Xpose database
}
\details{ 
\itemize{
\item \code{\link{add.absval}}
\item \code{\link{add.dichot}}
\item \code{\link{add.exp}}
\item \code{\link{add.log}}
\item \code{\link{add.tad}}
\item \code{\link{bootscm.import}}
\item \code{\link{calc.npar}}
\item \code{\link{change.ab.graph.par}}
\item \code{\link{change.bw.graph.par}}
\item \code{\link{change.cat.cont}}
\item \code{\link{change.cat.levels}}
\item \code{\link{change.cond.graph.par}}
\item \code{\link{change.dil.graph.par}}
\item \code{\link{change.dv.cat.levels}}
\item \code{\link{change.label.par}}
\item \code{\link{change.lm.graph.par}}
\item \code{\link{change.misc.graph.par}}
\item \code{\link{change.miss}}
\item \code{\link{change.parm}}
\item \code{\link{change.pi.graph.par}}
\item \code{\link{change.smooth.graph.par}}
\item \code{\link{change.subset}}
\item \code{\link{change.var.name}}
\item \code{\link{change.xlabel}}
\item \code{\link{change.xvardef}}
\item \code{\link{check.vars}}
\item \code{\link{create.parameter.list}}
\item \code{\link{createXposeClasses}}
\item \code{\link{Data}}
\item \code{\link{db.names}}
\item \code{\link{export.graph.par}}
\item \code{\link{export.variable.definitions}}
\item \code{\link{fix.wrapped.lines}}
\item \code{\link{get.doc}}
\item \code{\link{import.graph.par}}
\item \code{\link{import.variable.definitions}}
\item \code{\link{is.readable.file}}
\item \code{\link{make.sb.data}}
\item \code{\link{nsim}}
\item \code{\link{nsim<-}}
\item \code{\link{read.lst}}
\item \code{\link{read.lst6}}
\item \code{\link{read.lst7}}
\item \code{\link{read.nm.tables}}
\item \code{\link{read.npc.vpc.results}}
\item \code{\link{read.phi}}
\item \code{\link{read.TTE.sim.data}}
\item \code{\link{read.vpctab}}
\item \code{\link{reset.graph.par}}
\item \code{\link{reset.variables}}
\item \code{\link{SData}}
\item \code{\link{set.doc}}
\item \code{\link{simpraz.xpdb}}
\item \code{\link{simprazExample}}
\item \code{\link{test.xpose.data}}
\item \code{\link{xlabel}}
\item \code{\link{xpose.ask.for.filename}}
\item \code{\link{xpose.ask.for.lst}}
\item \code{\link{xpose.ask.for.mod}}
\item \code{\link{xpose.bin}}
\item \code{\link{xpose.data}}
\item \code{\link{xpose.data-class}}
\item \code{\link{xpose.prefs-class}}
\item \code{\link{xpose.print}}
\item \code{\link{xpose.read}}
\item \code{\link{xpose.write}}
\item \code{\link{xsubset}}
\item \code{\link{xvardef}}
}
}

\author{E. Niclas Jonsson, Mats Karlsson, Justin Wilkins and Andrew Hooker}
\keyword{methods}
\keyword{package}



